Get a list of Chemical Categories

Creates a list of chemical categories included in the database. Note this is not comprehensive.

get.categories(ver = "upd")

Arguments

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing the categories for all chemicals included the database.

Examples

get.categories()

#>             Category
#> 1                DDT
#> 2                PAH
#> 5         Haloalkane
#> 7                HCH
#> 8              Ether
#> 9            Alcohol
#> 10   Carboxylic Acid
#> 11          Aldehyde
#> 14            Ketone
#> 15          Haloform
#> 16      Organosulfur
#> 17             Amide
#> 21             Arene
#> 23       Cyclodienes
#> 26            Alkane
#> 28            Alkene
#> 31             Amine
#> 36           Nitrile
#> 40             Nitro
#> 43   Organophosphate
#> 54           Acetate
#> 57         Phthalate
#> 60         Haloarene
#> 63         Carbozole
#> 64                CB
#> 68               PBB
#> 78             Ester
#> 84           Formate
#> 90            Cyclic
#> 108      Piperidine 
#> 123      Pyrrolidine
#> 129          Dioxane
#> 133        Inorganic
#> 148       Haloalkene
#> 168      Organometal
#> 175    Fluorotelomer
#> 187    Organosilicon
#> 190            PFASs
#> 191              PCN
#> 193             PBDE
#> 194              PCB
#> 252        Cinnamate
#> 311          Aniline
#> 323          Terpene
#> 332         Biphenyl
#> 340          Paraben
#> 348    Diarylethene 
#> 355      Haloalcohol
#> 364         Pyridine
#> 383              CDD
#> 432              CDF
#> 966          Element
#> 975              CFC
#> 981         Labelled
#> 1028     Carboxylate
#> 1034          Phenol
#> 1068       Haloether
#> 1069 Carboxylic acid
#> 1070        Xanthine
#> 1072        Pyrazole
#> 1073         Steroid
#> 1076            Urea
#> 1077      Salicylate
#> 1080    Benzoic Acid
#> 1082   Prostaglandin
#> 1089           Azine
#> 1090       Diazapine
#> 1092   Proionic acid
#> 1113      Halocyclic
#> 1125     Barbiturate
#> 1131      Carboprost
#> 1135              BB
#> 1161            PCDE
#> 1271            musk
#> 1274            Musk
#> 1288         Camphor
#> 1417          Alkyne
#> 1511             BPA
#> 1557            FTOH
#> 1589      Morpholine
#> 1591      Piperazine
#> 1624    Benzaldehyde
#> 1659          Purine
#> 1674         Pyrrole
#> 1682       Haloester
#> 1695     Phthalimide
#> 1711   Halobisphenol
#> 1730     Isocyanuric
#> 1733  Halobenzofuran