Get a list of Citations

Creates a list of citations included in the database.

get.citations(ver = "upd")

Arguments

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing citations for all references included the database.

Examples

get.citations()

#>                             Citation
#> 1             Harner and Mackay 1995
#> 2           Harner and Bidleman 1996
#> 3                Finizio et al. 1997
#> 4            Kömp and McLachlan 1997
#> 5                 Gruber et al. 1997
#> 6                         Hiatt 1997
#> 7             Shoeib and Harner 2002
#> 8           Harner and Bidleman 1998
#> 9                  Zhang et al. 1999
#> 10                Harner et al. 2000
#> 11                  Chen et al. 2001
#> 12                     Ferreira 2001
#> 13               Abraham et al. 2001
#> 14                Treves et al. 2001
#> 15                 Chen et al. 2002a
#> 16                 Chen et al. 2002b
#> 17            Harner and Shoeib 2002
#> 18                    Su et al. 2002
#> 19                 Wania et al. 2002
#> 20                 Chen et al. 2003b
#> 21                 Chen et al. 2003c
#> 22                   Lei et al. 2004
#> 23                    Batterman 2002
#> 24                Shoeib et al. 2004
#> 25                Parnis et al. 2015
#> 26            Meylan and Howard 2005
#> 27          Puzyn and Falandysz 2005
#> 28                  Zhao et al. 2005
#> 29                   Li et al. 2006a
#> 30                  Goss et al. 2006
#> 31              Odabasi et al. 2006a
#> 32              Odabasi et al. 2006b
#> 33               Raevsky et al. 2006
#> 34                    Xu et al. 2007
#> 35                   Papa et a. 2009
#> 36        Sepassi and Yalkowsky 2007
#> 37                  Wang et al. 2008
#> 38                Thuens et al. 2008
#> 39                    Fu et al. 2016
#> 40                Taheri et al. 1991
#> 41                Dreyer et al. 2009
#> 42               Hongxia et al. 2009
#> 43                 Zhang et al. 2009
#> 44                  Ha and Kwon 2010
#> 45                  Zhao et al. 2010
#> 46            Odabasi and Cetin 2012
#> 47             Xu and Kropscott 2012
#> 48                   Liu et al. 2013
#> 49                  Zeng et al. 2013
#> 50             Xu and Kropscott 2013
#> 51                  Jiao et al. 2014
#> 52             Xu and Kropscott 2014
#> 53              Pegoraro et al. 2015
#> 54                  Chen et al. 2016
#> 55                  Yuan et al. 2016
#> 56                 Lee and Kwon 2016
#> 57           Shoeib and Harner 2002b
#> 58                  Wang et al. 2017
#> 59                   Jin et al. 2017
#> 60                   Lei et al. 2019
#> 61                 Chen et al. 2003a
#> 62  Unpublished data from Tom Harner
#> 63                  Fang et al. 1997
#> 64              Tse and Sandler 1994
#> 65                 Berti et al. 1986
#> 66                Cabani et al. 1991
#> 67                Thomas et al. 1982
#> 68                Tewari et al. 1982
#> 69                    Bo et al. 1993
#> 70            Boyer and Bircher 1960
#> 71               Wilcock et al. 1978
#> 72           Bhatia and Sandler 1995
#> 73                Raines et al. 2004
#> 74              Hussam and Carr 1985
#> 75                Rohrschneider 1973
#> 76                  Park et al. 1987
#> 77                       Dallas 1995
#> 78            Nedyalkova et al. 2019
#> 79       Abraham and Nasehzadeh 1981
#> 80              Dallas and Carr 1992
#> 81               Fukuchi et al. 1999
#> 82               Fukuchi et al. 2001
#> 83              Castells et al. 2000
#> 84           Gmehling and Menke 2007
#> 85              Castells et al. 1999
#> 86                      Abraham 1982
#> 87                        Hiatt 1998
#> 88                      Gerrard 1973
#> 89             Hernandez et al. 1999
#> 90                      Gerrard 1972
#> 91                Martin et al. 1984
#> 92               Pollack et al. 1984
#> 93            Thomas and Eckert 1984
#> 94                 Acree et al. 1994
#> 95             Beerbower et al. 1984
#> 96               De Fina et al. 2000
#> 97             Dohnal and Vrkba 1997
#> 98                Miller et al. 1985
#> 99              Pinsuwan et al. 1995
#> 100                      Eikens 1993
#> 101              Gardner et al. 1971
#> 102          Ellgehausen et al. 1981
#> 103            Yalkowsky et al. 1983
#> 104                Gobas et al. 1988
#> 105             Fletcher et al. 1997
#> 106               Powell et al. 1997
#> 107                       Niimi 1991
#> 108               Ruelle et al. 1994
#> 109       Granberg and Rasmuson 1999
#> 110                Chiou et al. 1982
#> 111               Powell et al. 1994
#> 112                      Ramble 2019
#> 113             Pierotti et al. 1959
#> 114                      Maaßen 1996
#> 115                        Kohn 1988
#> 116                      Cheong 1989
#> 117              Ruelle et al. 1994b
#> 118               Mackay et al. 1980
#> 119           Anliker and Moser 1987
#> 120               Ruelle et al. 1992
#> 121            Yalkowsky et al. 1994
#> 122        Kurz and Ballchmiter 1999
#> 123                   Li et al. 1995
#> 124       Sepassi and Yalkowsky 2006
#> 125          Lian and Yalkowsky 2014
#> 126                      US EPA 2012
#> 127                        Yaws 2012
#> 128                Okeme et al. 2020
#> 129                       Özcan 2013
#> 130                Vuong et al. 2020
#> 131               Taheri et al. 1993
#> 132              Ionescu et al. 1994
#> 133                 Fang et al. 1996
#> 134                Fang et al. 1997b
#> 135                 Eger et al. 1997
#> 136                 Best et al. 1997
#> 137                 Eger et al. 1999
#> 138          Cousins and Mackay 2000
#> 139                 Eger et al. 2001
#> 140           Oliferenko et al. 2004
#> 141              Abraham et al. 2005
#> 142                  Yao et al. 2007
#> 143                 Nabi et al. 2014
#> 144                 Leng et al. 2015
#> 145          Vikas and Chayawan 2015
#> 146                  Kim et al. 2016
#> 147                Yaman et al. 2020
#> 148               Giesen et al. 1997
#> 149                  Zhu et al. 1998
#> 150                   Li et al. 1999
#> 151         Duffy and Jorgensen 2000
#> 152                     Roberts 2005
#> 153                Zhang et al. 2016
#> 154                     Mathieu 2020
#> 155                   Li et al. 2020