Get a list of Groups

Creates a list of research groups included in the database.

get.groups(ver = "upd")

Arguments

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing research groups included the database.

Examples

get.groups()

#>              Team
#> 1          Harner
#> 4            Kömp
#> 5        Gmehling
#> 6           Hiatt
#> 9         Schramm
#> 11           Chen
#> 12       Ferreira
#> 13        Abraham
#> 14         Treves
#> 18          Wania
#> 23      Batterman
#> 26         Meylan
#> 27          Puzyn
#> 28          Liang
#> 31        Odabasi
#> 33        Raevsky
#> 34             Xu
#> 35      Gramatica
#> 36        Sepassi
#> 37           Zeng
#> 38         Dreyer
#> 40           Eger
#> 43          Zhang
#> 44           Kwon
#> 47      Kropscott
#> 48            Liu
#> 51           Jiao
#> 53      Chiappero
#> 54             Li
#> 55           Yuan
#> 64        Sandler
#> 65         Cabani
#> 67         Thomas
#> 68         Tewari
#> 69        Battino
#> 70          Boyer
#> 73         Raines
#> 74           Carr
#> 75  Rohrschneider
#> 78     Nedyalkova
#> 81        Fukuchi
#> 88        Gerrard
#> 89          Acree
#> 91         Martin
#> 92        Pollack
#> 97         Dohnal
#> 98          Wasik
#> 99      Yalkowsky
#> 101       Gardner
#> 102   Ellgehausen
#> 104        Mackay
#> 107         Niimi
#> 108        Ruelle
#> 109      Granberg
#> 110         Chiou
#> 112        Ramble
#> 113      Pierotti
#> 114        Maaßen
#> 115          Kohn
#> 119       Anliker
#> 122          Kurz
#> 126        US EPA
#> 127          Yaws
#> 128      Bidleman
#> 129         Ozcan
#> 130          Choi
#> 136          Best
#> 140    Oliferenko
#> 143          Nabi
#> 144          Leng
#> 145         Vikas
#> 146           Kim
#> 148       Truhlar
#> 151         Duffy
#> 152       Roberts
#> 153       Diamond
#> 154       Mathieu
#> 155   Peijnenburg