Search by CAS number

Create a data frame with all estimated and experimental KOA values which match the specified CAS number(s).

query.cas(query, more.info = F, ver = "upd")

Arguments

query
CAS number, with or without dashes, either a single input or a vector.

more.info
When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing KOA data for chemicals with CAS No. matching the query.

See also

Examples

query.cas("50-29-3")

#>     Cas_No                   Chemical_Name Temp   log_KOA Ranking Type
#> 1  50-29-3 Dichlorodiphenyltrichloroethane  -10 11.834221       4  EXP
#> 2  50-29-3 Dichlorodiphenyltrichloroethane    0 11.283516       4  EXP
#> 3  50-29-3 Dichlorodiphenyltrichloroethane   10 10.771731       4  EXP
#> 4  50-29-3 Dichlorodiphenyltrichloroethane   10 10.970000       3  EST
#> 5  50-29-3 Dichlorodiphenyltrichloroethane   15 10.306539       4  EXP
#> 6  50-29-3 Dichlorodiphenyltrichloroethane   15 10.749532       3  EST
#> 7  50-29-3 Dichlorodiphenyltrichloroethane   20 10.294881       4  EXP
#> 8  50-29-3 Dichlorodiphenyltrichloroethane   25  7.900000       3  EST
#> 9  50-29-3 Dichlorodiphenyltrichloroethane   25  9.337337       2  EST
#> 10 50-29-3 Dichlorodiphenyltrichloroethane   25  9.706624       3  EST
#> 11 50-29-3 Dichlorodiphenyltrichloroethane   25  9.816309       4  EXP
#> 12 50-29-3 Dichlorodiphenyltrichloroethane   25  9.878767       4  EXP
#> 13 50-29-3 Dichlorodiphenyltrichloroethane   25  9.930000       1  EST
#> 14 50-29-3 Dichlorodiphenyltrichloroethane   25 10.090000       4  EXP
#> 15 50-29-3 Dichlorodiphenyltrichloroethane   25 10.110000       3  EST
#> 16 50-29-3 Dichlorodiphenyltrichloroethane   25 10.380000       3  EST
#> 17 50-29-3 Dichlorodiphenyltrichloroethane   25 10.388991       3  EST
#> 18 50-29-3 Dichlorodiphenyltrichloroethane   35  9.295065       4  EXP
#> 19 50-29-3 Dichlorodiphenyltrichloroethane   35 10.051850       3  EST
#> 20 50-29-3 Dichlorodiphenyltrichloroethane   45  8.824000       4  EXP
#> 21 50-29-3 Dichlorodiphenyltrichloroethane   45  9.735903       3  EST
#>    Method_Bin                Citation
#> 1          FM  Harner and Mackay 1995
#> 2          FM  Harner and Mackay 1995
#> 3          FM  Harner and Mackay 1995
#> 4        QSPR  Meylan and Howard 2005
#> 5          FM Shoeib and Harner 2002b
#> 6        QSPR            Mathieu 2020
#> 7          FM  Harner and Mackay 1995
#> 8   Solvation          Fu et al. 2016
#> 9        QSPR        Nabi et al. 2014
#> 10       QSPR        Nabi et al. 2014
#> 11         FM Shoeib and Harner 2002b
#> 12         FM Shoeib and Harner 2002b
#> 13       QSPR     Finizio et al. 1997
#> 14         FM  Harner and Mackay 1995
#> 15       QSPR       Chen et al. 2003c
#> 16       QSPR  Meylan and Howard 2005
#> 17       QSPR            Mathieu 2020
#> 18         FM Shoeib and Harner 2002b
#> 19       QSPR            Mathieu 2020
#> 20         FM Shoeib and Harner 2002b
#> 21       QSPR            Mathieu 2020
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                  Reported in paper as recommended value
#> 12 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                Raw data and errors from communication with T. Harner. 
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11 Calculated from regression equation and originally reported in paper.
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                          Raw data from communication with T. Harner. 
#> 21                                                                  <NA>

query.cas("50293")

#>     Cas_No                   Chemical_Name Temp   log_KOA Ranking Type
#> 1  50-29-3 Dichlorodiphenyltrichloroethane  -10 11.834221       4  EXP
#> 2  50-29-3 Dichlorodiphenyltrichloroethane    0 11.283516       4  EXP
#> 3  50-29-3 Dichlorodiphenyltrichloroethane   10 10.771731       4  EXP
#> 4  50-29-3 Dichlorodiphenyltrichloroethane   10 10.970000       3  EST
#> 5  50-29-3 Dichlorodiphenyltrichloroethane   15 10.306539       4  EXP
#> 6  50-29-3 Dichlorodiphenyltrichloroethane   15 10.749532       3  EST
#> 7  50-29-3 Dichlorodiphenyltrichloroethane   20 10.294881       4  EXP
#> 8  50-29-3 Dichlorodiphenyltrichloroethane   25  7.900000       3  EST
#> 9  50-29-3 Dichlorodiphenyltrichloroethane   25  9.337337       2  EST
#> 10 50-29-3 Dichlorodiphenyltrichloroethane   25  9.706624       3  EST
#> 11 50-29-3 Dichlorodiphenyltrichloroethane   25  9.816309       4  EXP
#> 12 50-29-3 Dichlorodiphenyltrichloroethane   25  9.878767       4  EXP
#> 13 50-29-3 Dichlorodiphenyltrichloroethane   25  9.930000       1  EST
#> 14 50-29-3 Dichlorodiphenyltrichloroethane   25 10.090000       4  EXP
#> 15 50-29-3 Dichlorodiphenyltrichloroethane   25 10.110000       3  EST
#> 16 50-29-3 Dichlorodiphenyltrichloroethane   25 10.380000       3  EST
#> 17 50-29-3 Dichlorodiphenyltrichloroethane   25 10.388991       3  EST
#> 18 50-29-3 Dichlorodiphenyltrichloroethane   35  9.295065       4  EXP
#> 19 50-29-3 Dichlorodiphenyltrichloroethane   35 10.051850       3  EST
#> 20 50-29-3 Dichlorodiphenyltrichloroethane   45  8.824000       4  EXP
#> 21 50-29-3 Dichlorodiphenyltrichloroethane   45  9.735903       3  EST
#>    Method_Bin                Citation
#> 1          FM  Harner and Mackay 1995
#> 2          FM  Harner and Mackay 1995
#> 3          FM  Harner and Mackay 1995
#> 4        QSPR  Meylan and Howard 2005
#> 5          FM Shoeib and Harner 2002b
#> 6        QSPR            Mathieu 2020
#> 7          FM  Harner and Mackay 1995
#> 8   Solvation          Fu et al. 2016
#> 9        QSPR        Nabi et al. 2014
#> 10       QSPR        Nabi et al. 2014
#> 11         FM Shoeib and Harner 2002b
#> 12         FM Shoeib and Harner 2002b
#> 13       QSPR     Finizio et al. 1997
#> 14         FM  Harner and Mackay 1995
#> 15       QSPR       Chen et al. 2003c
#> 16       QSPR  Meylan and Howard 2005
#> 17       QSPR            Mathieu 2020
#> 18         FM Shoeib and Harner 2002b
#> 19       QSPR            Mathieu 2020
#> 20         FM Shoeib and Harner 2002b
#> 21       QSPR            Mathieu 2020
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                  Reported in paper as recommended value
#> 12 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                Raw data and errors from communication with T. Harner. 
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11 Calculated from regression equation and originally reported in paper.
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                          Raw data from communication with T. Harner. 
#> 21                                                                  <NA>

query.cas("50293", more.info = TRUE)

#>    Category  Cas_No                   Chemical_Name
#> 1       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 2       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 3       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 4       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 5       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 6       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 7       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 8       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 9       DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 10      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 11      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 12      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 13      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 14      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 15      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 16      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 17      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 18      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 19      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 20      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#> 21      DDT 50-29-3 Dichlorodiphenyltrichloroethane
#>                      Chemical_Name
#> 1  Dichlorodiphenyltrichloroethane
#> 2  Dichlorodiphenyltrichloroethane
#> 3  Dichlorodiphenyltrichloroethane
#> 4  Dichlorodiphenyltrichloroethane
#> 5  Dichlorodiphenyltrichloroethane
#> 6  Dichlorodiphenyltrichloroethane
#> 7  Dichlorodiphenyltrichloroethane
#> 8  Dichlorodiphenyltrichloroethane
#> 9  Dichlorodiphenyltrichloroethane
#> 10 Dichlorodiphenyltrichloroethane
#> 11 Dichlorodiphenyltrichloroethane
#> 12 Dichlorodiphenyltrichloroethane
#> 13 Dichlorodiphenyltrichloroethane
#> 14 Dichlorodiphenyltrichloroethane
#> 15 Dichlorodiphenyltrichloroethane
#> 16 Dichlorodiphenyltrichloroethane
#> 17 Dichlorodiphenyltrichloroethane
#> 18 Dichlorodiphenyltrichloroethane
#> 19 Dichlorodiphenyltrichloroethane
#> 20 Dichlorodiphenyltrichloroethane
#> 21 Dichlorodiphenyltrichloroethane
#>                                                   IUPAC_name
#> 1  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 2  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 3  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 4  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 5  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 6  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 7  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 8  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 9  1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 10 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 11 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 12 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 13 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 14 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 15 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 16 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 17 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 18 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 19 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 20 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#> 21 1,1'-(2,2,2-Trichloroethane-1,1-diyl)bis(4-chlorobenzene)
#>                                          Alt_Chem_Name  Acronym Alt_Acronym
#> 1  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 2  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 3  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 4  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 5  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 6  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 7  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 8  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 9  1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 10 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 11 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 12 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 13 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 14 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 15 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 16 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 17 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 18 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 19 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 20 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#> 21 1,1'-(2,2,2-trichloroethylidene)bis 4-chlorobenzene p,p'-DDT         DDT
#>    Molar_Mass Temp   log_KOA Ranking Type Method_Type            Gen_Method
#> 1      354.48  -10 11.834221       4  EXP     Dynamic                    FM
#> 2      354.48    0 11.283516       4  EXP     Dynamic                    FM
#> 3      354.48   10 10.771731       4  EXP     Dynamic                    FM
#> 4      354.48   10 10.970000       3  EST         EST              Triangle
#> 5      354.48   15 10.306539       4  EXP     Dynamic                    FM
#> 6      354.48   15 10.749532       3  EST         EST              additive
#> 7      354.48   20 10.294881       4  EXP     Dynamic                    FM
#> 8      354.48   25  7.900000       3  EST         EST SM8AD/CM4/M05-2X/6-31
#> 9      354.48   25  9.337337       2  EST         EST                ppLFER
#> 10     354.48   25  9.706624       3  EST         EST                  LFER
#> 11     354.48   25  9.816309       4  EXP     Dynamic                    FM
#> 12     354.48   25  9.878767       4  EXP     Dynamic                    FM
#> 13     354.48   25  9.930000       1  EST         EST              Triangle
#> 14     354.48   25 10.090000       4  EXP     Dynamic                    FM
#> 15     354.48   25 10.110000       3  EST         EST                   PLS
#> 16     354.48   25 10.380000       3  EST         EST              Triangle
#> 17     354.48   25 10.388991       3  EST         EST              additive
#> 18     354.48   35  9.295065       4  EXP     Dynamic                    FM
#> 19     354.48   35 10.051850       3  EST         EST              additive
#> 20     354.48   45  8.824000       4  EXP     Dynamic                    FM
#> 21     354.48   45  9.735903       3  EST         EST              additive
#>    Method_Bin wet_dry_Octanol Value_Reported_As                Citation    Team
#> 1          FM     dry octanol           log KOA  Harner and Mackay 1995  Harner
#> 2          FM     dry octanol           log KOA  Harner and Mackay 1995  Harner
#> 3          FM     dry octanol           log KOA  Harner and Mackay 1995  Harner
#> 4        QSPR     wet octanol           log KOA  Meylan and Howard 2005  Meylan
#> 5          FM     dry octanol           log KOA Shoeib and Harner 2002b  Harner
#> 6        QSPR     dry octanol           log KOA            Mathieu 2020 Mathieu
#> 7          FM     dry octanol           log KOA  Harner and Mackay 1995  Harner
#> 8   Solvation     dry octanol           log KOA          Fu et al. 2016    Chen
#> 9        QSPR     dry octanol           log KOA        Nabi et al. 2014    Nabi
#> 10       QSPR     dry octanol           log KOA        Nabi et al. 2014    Nabi
#> 11         FM     dry octanol           log KOA Shoeib and Harner 2002b  Harner
#> 12         FM     dry octanol           log KOA Shoeib and Harner 2002b  Harner
#> 13       QSPR     wet octanol           log KOA     Finizio et al. 1997  Harner
#> 14         FM     dry octanol           log KOA  Harner and Mackay 1995  Harner
#> 15       QSPR     dry octanol           log KOA       Chen et al. 2003c    Chen
#> 16       QSPR     wet octanol           log KOA  Meylan and Howard 2005  Meylan
#> 17       QSPR     dry octanol           log KOA            Mathieu 2020 Mathieu
#> 18         FM     dry octanol           log KOA Shoeib and Harner 2002b  Harner
#> 19       QSPR     dry octanol           log KOA            Mathieu 2020 Mathieu
#> 20         FM     dry octanol           log KOA Shoeib and Harner 2002b  Harner
#> 21       QSPR     dry octanol           log KOA            Mathieu 2020 Mathieu
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                  Reported in paper as recommended value
#> 12 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                Raw data and errors from communication with T. Harner. 
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11 Calculated from regression equation and originally reported in paper.
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                          Raw data from communication with T. Harner. 
#> 21                                                                  <NA>

query.cas(c("50-29-3","118-74-1"))

#>      Cas_No                   Chemical_Name Temp   log_KOA Ranking Type
#> 1   50-29-3 Dichlorodiphenyltrichloroethane  -10 11.834221       4  EXP
#> 2   50-29-3 Dichlorodiphenyltrichloroethane    0 11.283516       4  EXP
#> 3   50-29-3 Dichlorodiphenyltrichloroethane   10 10.771731       4  EXP
#> 4   50-29-3 Dichlorodiphenyltrichloroethane   10 10.970000       3  EST
#> 5   50-29-3 Dichlorodiphenyltrichloroethane   15 10.306539       4  EXP
#> 6   50-29-3 Dichlorodiphenyltrichloroethane   15 10.749532       3  EST
#> 7   50-29-3 Dichlorodiphenyltrichloroethane   20 10.294881       4  EXP
#> 8   50-29-3 Dichlorodiphenyltrichloroethane   25  7.900000       3  EST
#> 9   50-29-3 Dichlorodiphenyltrichloroethane   25  9.337337       2  EST
#> 10  50-29-3 Dichlorodiphenyltrichloroethane   25  9.706624       3  EST
#> 11  50-29-3 Dichlorodiphenyltrichloroethane   25  9.816309       4  EXP
#> 12  50-29-3 Dichlorodiphenyltrichloroethane   25  9.878767       4  EXP
#> 13  50-29-3 Dichlorodiphenyltrichloroethane   25  9.930000       1  EST
#> 14  50-29-3 Dichlorodiphenyltrichloroethane   25 10.090000       4  EXP
#> 15  50-29-3 Dichlorodiphenyltrichloroethane   25 10.110000       3  EST
#> 16  50-29-3 Dichlorodiphenyltrichloroethane   25 10.380000       3  EST
#> 17  50-29-3 Dichlorodiphenyltrichloroethane   25 10.388991       3  EST
#> 18  50-29-3 Dichlorodiphenyltrichloroethane   35  9.295065       4  EXP
#> 19  50-29-3 Dichlorodiphenyltrichloroethane   35 10.051850       3  EST
#> 20  50-29-3 Dichlorodiphenyltrichloroethane   45  8.824000       4  EXP
#> 21  50-29-3 Dichlorodiphenyltrichloroethane   45  9.735903       3  EST
#> 22 118-74-1               Hexachlorobenzene  -10  8.635361       4  EXP
#> 23 118-74-1               Hexachlorobenzene    0  8.088278       4  EXP
#> 24 118-74-1               Hexachlorobenzene    5  7.912302       3  EST
#> 25 118-74-1               Hexachlorobenzene    5  8.104375       4  EXP
#> 26 118-74-1               Hexachlorobenzene   10  7.579859       4  EXP
#> 27 118-74-1               Hexachlorobenzene   10  7.772583       3  EST
#> 28 118-74-1               Hexachlorobenzene   10  7.810000       3  EST
#> 29 118-74-1               Hexachlorobenzene   10  7.850000       3  EST
#> 30 118-74-1               Hexachlorobenzene   10  7.887005       4  EXP
#> 31 118-74-1               Hexachlorobenzene   10  7.998000       3  EST
#> 32 118-74-1               Hexachlorobenzene   15  7.637712       3  EST
#> 33 118-74-1               Hexachlorobenzene   15  7.699948       4  EXP
#> 34 118-74-1               Hexachlorobenzene   20  7.106143       4  EXP
#> 35 118-74-1               Hexachlorobenzene   20  7.330000       3  EST
#> 36 118-74-1               Hexachlorobenzene   20  7.507443       3  EST
#> 37 118-74-1               Hexachlorobenzene   20  7.563416       4  EXP
#> 38 118-74-1               Hexachlorobenzene   20  7.638000       3  EST
#> 39 118-74-1               Hexachlorobenzene   25  4.930000       3  EST
#> 40 118-74-1               Hexachlorobenzene   25  5.760000       3  EST
#> 41 118-74-1               Hexachlorobenzene   25  6.890000       3  EST
#> 42 118-74-1               Hexachlorobenzene   25  6.900000       4  EXP
#> 43 118-74-1               Hexachlorobenzene   25  6.910000       3  EST
#> 44 118-74-1               Hexachlorobenzene   25  6.910000       1  EST
#> 45 118-74-1               Hexachlorobenzene   25  7.132230       3  EST
#> 46 118-74-1               Hexachlorobenzene   25  7.170000       2  EST
#> 47 118-74-1               Hexachlorobenzene   25  7.324670       2  EST
#> 48 118-74-1               Hexachlorobenzene   25  7.381542       3  EST
#> 49 118-74-1               Hexachlorobenzene   25  7.383523       4  EXP
#> 50 118-74-1               Hexachlorobenzene   25  7.387799       4  EXP
#> 51 118-74-1               Hexachlorobenzene   25  7.440000       3  EST
#> 52 118-74-1               Hexachlorobenzene   25  7.880000       3  EST
#> 53 118-74-1               Hexachlorobenzene   25  8.540000       1  EST
#>    Method_Bin                   Citation
#> 1          FM     Harner and Mackay 1995
#> 2          FM     Harner and Mackay 1995
#> 3          FM     Harner and Mackay 1995
#> 4        QSPR     Meylan and Howard 2005
#> 5          FM    Shoeib and Harner 2002b
#> 6        QSPR               Mathieu 2020
#> 7          FM     Harner and Mackay 1995
#> 8   Solvation             Fu et al. 2016
#> 9        QSPR           Nabi et al. 2014
#> 10       QSPR           Nabi et al. 2014
#> 11         FM    Shoeib and Harner 2002b
#> 12         FM    Shoeib and Harner 2002b
#> 13       QSPR        Finizio et al. 1997
#> 14         FM     Harner and Mackay 1995
#> 15       QSPR          Chen et al. 2003c
#> 16       QSPR     Meylan and Howard 2005
#> 17       QSPR               Mathieu 2020
#> 18         FM    Shoeib and Harner 2002b
#> 19       QSPR               Mathieu 2020
#> 20         FM    Shoeib and Harner 2002b
#> 21       QSPR               Mathieu 2020
#> 22         FM     Harner and Mackay 1995
#> 23         FM     Harner and Mackay 1995
#> 24       QSPR               Mathieu 2020
#> 25         FM    Shoeib and Harner 2002b
#> 26         FM     Harner and Mackay 1995
#> 27       QSPR               Mathieu 2020
#> 28       QSPR            Jin et al. 2017
#> 29       QSPR     Meylan and Howard 2005
#> 30         FM    Shoeib and Harner 2002b
#> 31       QSPR            Li et al. 2006a
#> 32       QSPR               Mathieu 2020
#> 33         FM    Shoeib and Harner 2002b
#> 34         FM     Harner and Mackay 1995
#> 35       QSPR            Jin et al. 2017
#> 36       QSPR               Mathieu 2020
#> 37         FM    Shoeib and Harner 2002b
#> 38       QSPR            Li et al. 2006a
#> 39  Solvation             Fu et al. 2016
#> 40       QSPR      Yalkowsky et al. 1994
#> 41       QSPR     Meylan and Howard 2005
#> 42         FM     Harner and Mackay 1995
#> 43       QSPR          Chen et al. 2003c
#> 44       QSPR           Zhao et al. 2005
#> 45       QSPR           Nabi et al. 2014
#> 46       QSPR        Abraham et al. 2001
#> 47       QSPR           Nabi et al. 2014
#> 48       QSPR               Mathieu 2020
#> 49         FM    Shoeib and Harner 2002b
#> 50         FM    Shoeib and Harner 2002b
#> 51       QSPR            Li et al. 2006a
#> 52       QSPR Sepassi and Yalkowsky 2007
#> 53       QSPR        Finizio et al. 1997
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                  Reported in paper as recommended value
#> 12 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#> 22                                                                                                                                    <NA>
#> 23                                                                                                                                    <NA>
#> 24                                                                                                                                    <NA>
#> 25                                                                                                                                    <NA>
#> 26                                                                                                                                    <NA>
#> 27                                                                                                                                    <NA>
#> 28                                                                                                                                    <NA>
#> 29                                                                                                                                    <NA>
#> 30                                                                                                                                    <NA>
#> 31                                                                                                                                    <NA>
#> 32                                                                                                                                    <NA>
#> 33                                                                                                                                    <NA>
#> 34                                                                                                                                    <NA>
#> 35                                                                                                                                    <NA>
#> 36                                                                                                                                    <NA>
#> 37                                                                                                                                    <NA>
#> 38                                                                                                                                    <NA>
#> 39                                                                                                                                    <NA>
#> 40                                                                                                                                    <NA>
#> 41                                                                                                                                    <NA>
#> 42                                                                                                                                    <NA>
#> 43                                                                                                                                    <NA>
#> 44                                                                                                                                    <NA>
#> 45                                                                                                                                    <NA>
#> 46                                                                                                                                    <NA>
#> 47                                                                                                                                    <NA>
#> 48                                                                                                                                    <NA>
#> 49                                                                                                  Reported in paper as recommended value
#> 50 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 51                                                                                                                                    <NA>
#> 52                                                                                                                                    <NA>
#> 53                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                Raw data and errors from communication with T. Harner. 
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11 Calculated from regression equation and originally reported in paper.
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                          Raw data from communication with T. Harner. 
#> 21                                                                  <NA>
#> 22                                                                  <NA>
#> 23                                                                  <NA>
#> 24                                                                  <NA>
#> 25               Raw data and errors from communication with T. Harner. 
#> 26                                                                  <NA>
#> 27                                                                  <NA>
#> 28                                                                  <NA>
#> 29                                                                  <NA>
#> 30               Raw data and errors from communication with T. Harner. 
#> 31                                                                  <NA>
#> 32                                                                  <NA>
#> 33               Raw data and errors from communication with T. Harner. 
#> 34                                                                  <NA>
#> 35                                                                  <NA>
#> 36                                                                  <NA>
#> 37               Raw data and errors from communication with T. Harner. 
#> 38                                                                  <NA>
#> 39                                                                  <NA>
#> 40                                                                  <NA>
#> 41                                                                  <NA>
#> 42                                                                  <NA>
#> 43                                                                  <NA>
#> 44                                                                  <NA>
#> 45                                                                  <NA>
#> 46                                                                  <NA>
#> 47                                                                  <NA>
#> 48                                                                  <NA>
#> 49 Calculated from regression equation and originally reported in paper.
#> 50               Raw data and errors from communication with T. Harner. 
#> 51                                                                  <NA>
#> 52                                                                  <NA>
#> 53                                                                  <NA>

query.cas("50293", ver = "upd")

#>     Cas_No                   Chemical_Name Temp   log_KOA Ranking Type
#> 1  50-29-3 Dichlorodiphenyltrichloroethane  -10 11.834221       4  EXP
#> 2  50-29-3 Dichlorodiphenyltrichloroethane    0 11.283516       4  EXP
#> 3  50-29-3 Dichlorodiphenyltrichloroethane   10 10.771731       4  EXP
#> 4  50-29-3 Dichlorodiphenyltrichloroethane   10 10.970000       3  EST
#> 5  50-29-3 Dichlorodiphenyltrichloroethane   15 10.306539       4  EXP
#> 6  50-29-3 Dichlorodiphenyltrichloroethane   15 10.749532       3  EST
#> 7  50-29-3 Dichlorodiphenyltrichloroethane   20 10.294881       4  EXP
#> 8  50-29-3 Dichlorodiphenyltrichloroethane   25  7.900000       3  EST
#> 9  50-29-3 Dichlorodiphenyltrichloroethane   25  9.337337       2  EST
#> 10 50-29-3 Dichlorodiphenyltrichloroethane   25  9.706624       3  EST
#> 11 50-29-3 Dichlorodiphenyltrichloroethane   25  9.816309       4  EXP
#> 12 50-29-3 Dichlorodiphenyltrichloroethane   25  9.878767       4  EXP
#> 13 50-29-3 Dichlorodiphenyltrichloroethane   25  9.930000       1  EST
#> 14 50-29-3 Dichlorodiphenyltrichloroethane   25 10.090000       4  EXP
#> 15 50-29-3 Dichlorodiphenyltrichloroethane   25 10.110000       3  EST
#> 16 50-29-3 Dichlorodiphenyltrichloroethane   25 10.380000       3  EST
#> 17 50-29-3 Dichlorodiphenyltrichloroethane   25 10.388991       3  EST
#> 18 50-29-3 Dichlorodiphenyltrichloroethane   35  9.295065       4  EXP
#> 19 50-29-3 Dichlorodiphenyltrichloroethane   35 10.051850       3  EST
#> 20 50-29-3 Dichlorodiphenyltrichloroethane   45  8.824000       4  EXP
#> 21 50-29-3 Dichlorodiphenyltrichloroethane   45  9.735903       3  EST
#>    Method_Bin                Citation
#> 1          FM  Harner and Mackay 1995
#> 2          FM  Harner and Mackay 1995
#> 3          FM  Harner and Mackay 1995
#> 4        QSPR  Meylan and Howard 2005
#> 5          FM Shoeib and Harner 2002b
#> 6        QSPR            Mathieu 2020
#> 7          FM  Harner and Mackay 1995
#> 8   Solvation          Fu et al. 2016
#> 9        QSPR        Nabi et al. 2014
#> 10       QSPR        Nabi et al. 2014
#> 11         FM Shoeib and Harner 2002b
#> 12         FM Shoeib and Harner 2002b
#> 13       QSPR     Finizio et al. 1997
#> 14         FM  Harner and Mackay 1995
#> 15       QSPR       Chen et al. 2003c
#> 16       QSPR  Meylan and Howard 2005
#> 17       QSPR            Mathieu 2020
#> 18         FM Shoeib and Harner 2002b
#> 19       QSPR            Mathieu 2020
#> 20         FM Shoeib and Harner 2002b
#> 21       QSPR            Mathieu 2020
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                  Reported in paper as recommended value
#> 12 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                Raw data and errors from communication with T. Harner. 
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11 Calculated from regression equation and originally reported in paper.
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                          Raw data from communication with T. Harner. 
#> 21                                                                  <NA>