Search by citation

Create a data frame with all estimated and experimental KOA values from the citations in the query.

query.citation(query, more.info = F, ver = "upd")

Arguments

query
Citation, either a single input or a vector. Use function ref.table() to get a list of citations in the database.

more.info
When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing KOA data from the citations included in the query.

See also

Examples

query.citation("Lei et al. 2019")

#>        Cas_No              Chemical_Name Temp  log_KOA Ranking Type Method_Bin
#> 1     67-64-1                    Acetone   25 2.320000       4  EXP         HS
#> 2     67-64-1                    Acetone   37 2.190000       4  EXP         HS
#> 3     67-64-1                    Acetone   40 2.167317       4  EXP         HS
#> 4     67-64-1                    Acetone   50 2.025306       4  EXP         HS
#> 5     67-64-1                    Acetone   60 2.000000       4  EXP         HS
#> 6     67-64-1                    Acetone   80 1.792392       4  EXP         HS
#> 7     67-64-1                    Acetone   90 1.724276       4  EXP         HS
#> 8     67-64-1                    Acetone  100 1.623249       4  EXP         HS
#> 9     71-23-8                   Propanol   25 3.820000       4  EXP         HS
#> 10    71-23-8                   Propanol   37 3.520000       4  EXP         HS
#> 11    71-23-8                   Propanol   40 3.445915       4  EXP         HS
#> 12    71-23-8                   Propanol   50 3.221675       4  EXP         HS
#> 13    71-23-8                   Propanol   60 3.005609       4  EXP         HS
#> 14    71-23-8                   Propanol   70 2.784617       4  EXP         HS
#> 15    71-23-8                   Propanol   80 2.587711       4  EXP         HS
#> 16    71-23-8                   Propanol   90 2.418301       4  EXP         HS
#> 17    71-23-8                   Propanol  100 2.267172       4  EXP         HS
#> 18    71-36-3                    Butanol   25 4.100000       4  EXP         HS
#> 19    71-36-3                    Butanol   37 3.810000       4  EXP         HS
#> 20    71-36-3                    Butanol   40 3.643946       4  EXP         HS
#> 21    71-36-3                    Butanol   50 3.410946       4  EXP         HS
#> 22    71-36-3                    Butanol   60 3.317227       4  EXP         HS
#> 23    71-36-3                    Butanol   70 3.118595       4  EXP         HS
#> 24    71-36-3                    Butanol   80 2.938019       4  EXP         HS
#> 25    71-36-3                    Butanol   90 2.781755       4  EXP         HS
#> 26    71-36-3                    Butanol  100 2.595496       4  EXP         HS
#> 27 1975-05-08               Acetonitrile   25 2.270000       4  EXP         HS
#> 28 1975-05-08               Acetonitrile   37 2.140000       4  EXP         HS
#> 29 1975-05-08               Acetonitrile   40 2.110590       4  EXP         HS
#> 30 1975-05-08               Acetonitrile   50 1.968483       4  EXP         HS
#> 31 1975-05-08               Acetonitrile   60 1.924279       4  EXP         HS
#> 32 1975-05-08               Acetonitrile   70 1.851258       4  EXP         HS
#> 33 1975-05-08               Acetonitrile   80 1.732394       4  EXP         HS
#> 34 1975-05-08               Acetonitrile   90 1.633468       4  EXP         HS
#> 35 1975-05-08               Acetonitrile  100 1.579784       4  EXP         HS
#> 36   106-42-3                   p-Xylene   25 3.590000       4  EXP         HS
#> 37   106-42-3                   p-Xylene   37 3.390000       4  EXP         HS
#> 38   106-42-3                   p-Xylene   70 2.885361       4  EXP         HS
#> 39   106-42-3                   p-Xylene   80 2.808211       4  EXP         HS
#> 40   106-42-3                   p-Xylene   90 2.673942       4  EXP         HS
#> 41   106-42-3                   p-Xylene  100 2.532754       4  EXP         HS
#> 42   106-42-3                   p-Xylene  110 2.457882       4  EXP         HS
#> 43   107-06-2         1,2-Dichloroethane   25 2.730000       4  EXP         HS
#> 44   107-06-2         1,2-Dichloroethane   37 2.570000       4  EXP         HS
#> 45   107-06-2         1,2-Dichloroethane   40 2.525045       4  EXP         HS
#> 46   107-06-2         1,2-Dichloroethane   50 2.380211       4  EXP         HS
#> 47   107-06-2         1,2-Dichloroethane   60 2.350248       4  EXP         HS
#> 48   107-06-2         1,2-Dichloroethane   70 2.206826       4  EXP         HS
#> 49   107-06-2         1,2-Dichloroethane   80 2.096910       4  EXP         HS
#> 50   107-06-2         1,2-Dichloroethane   90 1.973128       4  EXP         HS
#> 51   107-06-2         1,2-Dichloroethane  100 1.897627       4  EXP         HS
#> 52   141-78-6              Ethyl acetate   25 2.630000       4  EXP         HS
#> 53   141-78-6              Ethyl acetate   37 2.480000       4  EXP         HS
#> 54   141-78-6              Ethyl acetate   40 2.401401       4  EXP         HS
#> 55   141-78-6              Ethyl acetate   50 2.357935       4  EXP         HS
#> 56   141-78-6              Ethyl acetate   60 2.305351       4  EXP         HS
#> 57   141-78-6              Ethyl acetate   70 2.149219       4  EXP         HS
#> 58   141-78-6              Ethyl acetate   80 2.049218       4  EXP         HS
#> 59   141-78-6              Ethyl acetate   90 1.968483       4  EXP         HS
#> 60   141-78-6              Ethyl acetate  100 1.875061       4  EXP         HS
#> 61   541-73-1        1,3-Dichlorobenzene   25 4.170000       4  EXP         HS
#> 62   541-73-1        1,3-Dichlorobenzene   37 3.940000       4  EXP         HS
#> 63   541-73-1        1,3-Dichlorobenzene   80 3.247973       4  EXP         HS
#> 64   541-73-1        1,3-Dichlorobenzene   90 3.111263       4  EXP         HS
#> 65   541-73-1        1,3-Dichlorobenzene  100 2.992554       4  EXP         HS
#> 66   541-73-1        1,3-Dichlorobenzene  110 2.845718       4  EXP         HS
#> 67   493-02-7              trans-decalin   25 4.200000       4  EXP         HS
#> 68   493-02-7              trans-decalin   37 3.980000       4  EXP         HS
#> 69   493-02-7              trans-decalin   80 3.331225       4  EXP         HS
#> 70   493-02-7              trans-decalin   90 3.177825       4  EXP         HS
#> 71   493-02-7              trans-decalin  100 3.118926       4  EXP         HS
#> 72   493-02-7              trans-decalin  110 2.938019       4  EXP         HS
#> 73   488-23-3 1,2,3,4-Tetramethylbenzene   25 4.400000       4  EXP         HS
#> 74   488-23-3 1,2,3,4-Tetramethylbenzene   37 4.300000       4  EXP         HS
#> 75   488-23-3 1,2,3,4-Tetramethylbenzene   80 3.719331       4  EXP         HS
#> 76   488-23-3 1,2,3,4-Tetramethylbenzene   90 3.568554       4  EXP         HS
#> 77   488-23-3 1,2,3,4-Tetramethylbenzene  100 3.533009       4  EXP         HS
#> 78   488-23-3 1,2,3,4-Tetramethylbenzene  110 3.378761       4  EXP         HS
#>           Citation Reliability Comment
#> 1  Lei et al. 2019        <NA>    <NA>
#> 2  Lei et al. 2019        <NA>    <NA>
#> 3  Lei et al. 2019        <NA>    <NA>
#> 4  Lei et al. 2019        <NA>    <NA>
#> 5  Lei et al. 2019        <NA>    <NA>
#> 6  Lei et al. 2019        <NA>    <NA>
#> 7  Lei et al. 2019        <NA>    <NA>
#> 8  Lei et al. 2019        <NA>    <NA>
#> 9  Lei et al. 2019        <NA>    <NA>
#> 10 Lei et al. 2019        <NA>    <NA>
#> 11 Lei et al. 2019        <NA>    <NA>
#> 12 Lei et al. 2019        <NA>    <NA>
#> 13 Lei et al. 2019        <NA>    <NA>
#> 14 Lei et al. 2019        <NA>    <NA>
#> 15 Lei et al. 2019        <NA>    <NA>
#> 16 Lei et al. 2019        <NA>    <NA>
#> 17 Lei et al. 2019        <NA>    <NA>
#> 18 Lei et al. 2019        <NA>    <NA>
#> 19 Lei et al. 2019        <NA>    <NA>
#> 20 Lei et al. 2019        <NA>    <NA>
#> 21 Lei et al. 2019        <NA>    <NA>
#> 22 Lei et al. 2019        <NA>    <NA>
#> 23 Lei et al. 2019        <NA>    <NA>
#> 24 Lei et al. 2019        <NA>    <NA>
#> 25 Lei et al. 2019        <NA>    <NA>
#> 26 Lei et al. 2019        <NA>    <NA>
#> 27 Lei et al. 2019        <NA>    <NA>
#> 28 Lei et al. 2019        <NA>    <NA>
#> 29 Lei et al. 2019        <NA>    <NA>
#> 30 Lei et al. 2019        <NA>    <NA>
#> 31 Lei et al. 2019        <NA>    <NA>
#> 32 Lei et al. 2019        <NA>    <NA>
#> 33 Lei et al. 2019        <NA>    <NA>
#> 34 Lei et al. 2019        <NA>    <NA>
#> 35 Lei et al. 2019        <NA>    <NA>
#> 36 Lei et al. 2019        <NA>    <NA>
#> 37 Lei et al. 2019        <NA>    <NA>
#> 38 Lei et al. 2019        <NA>    <NA>
#> 39 Lei et al. 2019        <NA>    <NA>
#> 40 Lei et al. 2019        <NA>    <NA>
#> 41 Lei et al. 2019        <NA>    <NA>
#> 42 Lei et al. 2019        <NA>    <NA>
#> 43 Lei et al. 2019        <NA>    <NA>
#> 44 Lei et al. 2019        <NA>    <NA>
#> 45 Lei et al. 2019        <NA>    <NA>
#> 46 Lei et al. 2019        <NA>    <NA>
#> 47 Lei et al. 2019        <NA>    <NA>
#> 48 Lei et al. 2019        <NA>    <NA>
#> 49 Lei et al. 2019        <NA>    <NA>
#> 50 Lei et al. 2019        <NA>    <NA>
#> 51 Lei et al. 2019        <NA>    <NA>
#> 52 Lei et al. 2019        <NA>    <NA>
#> 53 Lei et al. 2019        <NA>    <NA>
#> 54 Lei et al. 2019        <NA>    <NA>
#> 55 Lei et al. 2019        <NA>    <NA>
#> 56 Lei et al. 2019        <NA>    <NA>
#> 57 Lei et al. 2019        <NA>    <NA>
#> 58 Lei et al. 2019        <NA>    <NA>
#> 59 Lei et al. 2019        <NA>    <NA>
#> 60 Lei et al. 2019        <NA>    <NA>
#> 61 Lei et al. 2019        <NA>    <NA>
#> 62 Lei et al. 2019        <NA>    <NA>
#> 63 Lei et al. 2019        <NA>    <NA>
#> 64 Lei et al. 2019        <NA>    <NA>
#> 65 Lei et al. 2019        <NA>    <NA>
#> 66 Lei et al. 2019        <NA>    <NA>
#> 67 Lei et al. 2019        <NA>    <NA>
#> 68 Lei et al. 2019        <NA>    <NA>
#> 69 Lei et al. 2019        <NA>    <NA>
#> 70 Lei et al. 2019        <NA>    <NA>
#> 71 Lei et al. 2019        <NA>    <NA>
#> 72 Lei et al. 2019        <NA>    <NA>
#> 73 Lei et al. 2019        <NA>    <NA>
#> 74 Lei et al. 2019        <NA>    <NA>
#> 75 Lei et al. 2019        <NA>    <NA>
#> 76 Lei et al. 2019        <NA>    <NA>
#> 77 Lei et al. 2019        <NA>    <NA>
#> 78 Lei et al. 2019        <NA>    <NA>

query.citation(c("Hussam and Carr 1985", "Lei et al. 2019"))

#>        Cas_No              Chemical_Name   Temp  log_KOA Ranking Type
#> 1     75-52-5               Nitromethane  25.01 2.584310       3  EXP
#> 2    108-88-3                    Toluene  25.01 3.292741       3  EXP
#> 3     67-64-1                    Acetone  25.00 2.320000       4  EXP
#> 4     67-64-1                    Acetone  37.00 2.190000       4  EXP
#> 5     67-64-1                    Acetone  40.00 2.167317       4  EXP
#> 6     67-64-1                    Acetone  50.00 2.025306       4  EXP
#> 7     67-64-1                    Acetone  60.00 2.000000       4  EXP
#> 8     67-64-1                    Acetone  80.00 1.792392       4  EXP
#> 9     67-64-1                    Acetone  90.00 1.724276       4  EXP
#> 10    67-64-1                    Acetone 100.00 1.623249       4  EXP
#> 11    71-23-8                   Propanol  25.00 3.820000       4  EXP
#> 12    71-23-8                   Propanol  37.00 3.520000       4  EXP
#> 13    71-23-8                   Propanol  40.00 3.445915       4  EXP
#> 14    71-23-8                   Propanol  50.00 3.221675       4  EXP
#> 15    71-23-8                   Propanol  60.00 3.005609       4  EXP
#> 16    71-23-8                   Propanol  70.00 2.784617       4  EXP
#> 17    71-23-8                   Propanol  80.00 2.587711       4  EXP
#> 18    71-23-8                   Propanol  90.00 2.418301       4  EXP
#> 19    71-23-8                   Propanol 100.00 2.267172       4  EXP
#> 20    71-36-3                    Butanol  25.00 4.100000       4  EXP
#> 21    71-36-3                    Butanol  37.00 3.810000       4  EXP
#> 22    71-36-3                    Butanol  40.00 3.643946       4  EXP
#> 23    71-36-3                    Butanol  50.00 3.410946       4  EXP
#> 24    71-36-3                    Butanol  60.00 3.317227       4  EXP
#> 25    71-36-3                    Butanol  70.00 3.118595       4  EXP
#> 26    71-36-3                    Butanol  80.00 2.938019       4  EXP
#> 27    71-36-3                    Butanol  90.00 2.781755       4  EXP
#> 28    71-36-3                    Butanol 100.00 2.595496       4  EXP
#> 29 1975-05-08               Acetonitrile  25.00 2.270000       4  EXP
#> 30 1975-05-08               Acetonitrile  37.00 2.140000       4  EXP
#> 31 1975-05-08               Acetonitrile  40.00 2.110590       4  EXP
#> 32 1975-05-08               Acetonitrile  50.00 1.968483       4  EXP
#> 33 1975-05-08               Acetonitrile  60.00 1.924279       4  EXP
#> 34 1975-05-08               Acetonitrile  70.00 1.851258       4  EXP
#> 35 1975-05-08               Acetonitrile  80.00 1.732394       4  EXP
#> 36 1975-05-08               Acetonitrile  90.00 1.633468       4  EXP
#> 37 1975-05-08               Acetonitrile 100.00 1.579784       4  EXP
#> 38   106-42-3                   p-Xylene  25.00 3.590000       4  EXP
#> 39   106-42-3                   p-Xylene  37.00 3.390000       4  EXP
#> 40   106-42-3                   p-Xylene  70.00 2.885361       4  EXP
#> 41   106-42-3                   p-Xylene  80.00 2.808211       4  EXP
#> 42   106-42-3                   p-Xylene  90.00 2.673942       4  EXP
#> 43   106-42-3                   p-Xylene 100.00 2.532754       4  EXP
#> 44   106-42-3                   p-Xylene 110.00 2.457882       4  EXP
#> 45   107-06-2         1,2-Dichloroethane  25.00 2.730000       4  EXP
#> 46   107-06-2         1,2-Dichloroethane  37.00 2.570000       4  EXP
#> 47   107-06-2         1,2-Dichloroethane  40.00 2.525045       4  EXP
#> 48   107-06-2         1,2-Dichloroethane  50.00 2.380211       4  EXP
#> 49   107-06-2         1,2-Dichloroethane  60.00 2.350248       4  EXP
#> 50   107-06-2         1,2-Dichloroethane  70.00 2.206826       4  EXP
#> 51   107-06-2         1,2-Dichloroethane  80.00 2.096910       4  EXP
#> 52   107-06-2         1,2-Dichloroethane  90.00 1.973128       4  EXP
#> 53   107-06-2         1,2-Dichloroethane 100.00 1.897627       4  EXP
#> 54   141-78-6              Ethyl acetate  25.00 2.630000       4  EXP
#> 55   141-78-6              Ethyl acetate  37.00 2.480000       4  EXP
#> 56   141-78-6              Ethyl acetate  40.00 2.401401       4  EXP
#> 57   141-78-6              Ethyl acetate  50.00 2.357935       4  EXP
#> 58   141-78-6              Ethyl acetate  60.00 2.305351       4  EXP
#> 59   141-78-6              Ethyl acetate  70.00 2.149219       4  EXP
#> 60   141-78-6              Ethyl acetate  80.00 2.049218       4  EXP
#> 61   141-78-6              Ethyl acetate  90.00 1.968483       4  EXP
#> 62   141-78-6              Ethyl acetate 100.00 1.875061       4  EXP
#> 63   541-73-1        1,3-Dichlorobenzene  25.00 4.170000       4  EXP
#> 64   541-73-1        1,3-Dichlorobenzene  37.00 3.940000       4  EXP
#> 65   541-73-1        1,3-Dichlorobenzene  80.00 3.247973       4  EXP
#> 66   541-73-1        1,3-Dichlorobenzene  90.00 3.111263       4  EXP
#> 67   541-73-1        1,3-Dichlorobenzene 100.00 2.992554       4  EXP
#> 68   541-73-1        1,3-Dichlorobenzene 110.00 2.845718       4  EXP
#> 69   493-02-7              trans-decalin  25.00 4.200000       4  EXP
#> 70   493-02-7              trans-decalin  37.00 3.980000       4  EXP
#> 71   493-02-7              trans-decalin  80.00 3.331225       4  EXP
#> 72   493-02-7              trans-decalin  90.00 3.177825       4  EXP
#> 73   493-02-7              trans-decalin 100.00 3.118926       4  EXP
#> 74   493-02-7              trans-decalin 110.00 2.938019       4  EXP
#> 75   488-23-3 1,2,3,4-Tetramethylbenzene  25.00 4.400000       4  EXP
#> 76   488-23-3 1,2,3,4-Tetramethylbenzene  37.00 4.300000       4  EXP
#> 77   488-23-3 1,2,3,4-Tetramethylbenzene  80.00 3.719331       4  EXP
#> 78   488-23-3 1,2,3,4-Tetramethylbenzene  90.00 3.568554       4  EXP
#> 79   488-23-3 1,2,3,4-Tetramethylbenzene 100.00 3.533009       4  EXP
#> 80   488-23-3 1,2,3,4-Tetramethylbenzene 110.00 3.378761       4  EXP
#>    Method_Bin             Citation Reliability Comment
#> 1          HS Hussam and Carr 1985        <NA>    <NA>
#> 2          HS Hussam and Carr 1985        <NA>    <NA>
#> 3          HS      Lei et al. 2019        <NA>    <NA>
#> 4          HS      Lei et al. 2019        <NA>    <NA>
#> 5          HS      Lei et al. 2019        <NA>    <NA>
#> 6          HS      Lei et al. 2019        <NA>    <NA>
#> 7          HS      Lei et al. 2019        <NA>    <NA>
#> 8          HS      Lei et al. 2019        <NA>    <NA>
#> 9          HS      Lei et al. 2019        <NA>    <NA>
#> 10         HS      Lei et al. 2019        <NA>    <NA>
#> 11         HS      Lei et al. 2019        <NA>    <NA>
#> 12         HS      Lei et al. 2019        <NA>    <NA>
#> 13         HS      Lei et al. 2019        <NA>    <NA>
#> 14         HS      Lei et al. 2019        <NA>    <NA>
#> 15         HS      Lei et al. 2019        <NA>    <NA>
#> 16         HS      Lei et al. 2019        <NA>    <NA>
#> 17         HS      Lei et al. 2019        <NA>    <NA>
#> 18         HS      Lei et al. 2019        <NA>    <NA>
#> 19         HS      Lei et al. 2019        <NA>    <NA>
#> 20         HS      Lei et al. 2019        <NA>    <NA>
#> 21         HS      Lei et al. 2019        <NA>    <NA>
#> 22         HS      Lei et al. 2019        <NA>    <NA>
#> 23         HS      Lei et al. 2019        <NA>    <NA>
#> 24         HS      Lei et al. 2019        <NA>    <NA>
#> 25         HS      Lei et al. 2019        <NA>    <NA>
#> 26         HS      Lei et al. 2019        <NA>    <NA>
#> 27         HS      Lei et al. 2019        <NA>    <NA>
#> 28         HS      Lei et al. 2019        <NA>    <NA>
#> 29         HS      Lei et al. 2019        <NA>    <NA>
#> 30         HS      Lei et al. 2019        <NA>    <NA>
#> 31         HS      Lei et al. 2019        <NA>    <NA>
#> 32         HS      Lei et al. 2019        <NA>    <NA>
#> 33         HS      Lei et al. 2019        <NA>    <NA>
#> 34         HS      Lei et al. 2019        <NA>    <NA>
#> 35         HS      Lei et al. 2019        <NA>    <NA>
#> 36         HS      Lei et al. 2019        <NA>    <NA>
#> 37         HS      Lei et al. 2019        <NA>    <NA>
#> 38         HS      Lei et al. 2019        <NA>    <NA>
#> 39         HS      Lei et al. 2019        <NA>    <NA>
#> 40         HS      Lei et al. 2019        <NA>    <NA>
#> 41         HS      Lei et al. 2019        <NA>    <NA>
#> 42         HS      Lei et al. 2019        <NA>    <NA>
#> 43         HS      Lei et al. 2019        <NA>    <NA>
#> 44         HS      Lei et al. 2019        <NA>    <NA>
#> 45         HS      Lei et al. 2019        <NA>    <NA>
#> 46         HS      Lei et al. 2019        <NA>    <NA>
#> 47         HS      Lei et al. 2019        <NA>    <NA>
#> 48         HS      Lei et al. 2019        <NA>    <NA>
#> 49         HS      Lei et al. 2019        <NA>    <NA>
#> 50         HS      Lei et al. 2019        <NA>    <NA>
#> 51         HS      Lei et al. 2019        <NA>    <NA>
#> 52         HS      Lei et al. 2019        <NA>    <NA>
#> 53         HS      Lei et al. 2019        <NA>    <NA>
#> 54         HS      Lei et al. 2019        <NA>    <NA>
#> 55         HS      Lei et al. 2019        <NA>    <NA>
#> 56         HS      Lei et al. 2019        <NA>    <NA>
#> 57         HS      Lei et al. 2019        <NA>    <NA>
#> 58         HS      Lei et al. 2019        <NA>    <NA>
#> 59         HS      Lei et al. 2019        <NA>    <NA>
#> 60         HS      Lei et al. 2019        <NA>    <NA>
#> 61         HS      Lei et al. 2019        <NA>    <NA>
#> 62         HS      Lei et al. 2019        <NA>    <NA>
#> 63         HS      Lei et al. 2019        <NA>    <NA>
#> 64         HS      Lei et al. 2019        <NA>    <NA>
#> 65         HS      Lei et al. 2019        <NA>    <NA>
#> 66         HS      Lei et al. 2019        <NA>    <NA>
#> 67         HS      Lei et al. 2019        <NA>    <NA>
#> 68         HS      Lei et al. 2019        <NA>    <NA>
#> 69         HS      Lei et al. 2019        <NA>    <NA>
#> 70         HS      Lei et al. 2019        <NA>    <NA>
#> 71         HS      Lei et al. 2019        <NA>    <NA>
#> 72         HS      Lei et al. 2019        <NA>    <NA>
#> 73         HS      Lei et al. 2019        <NA>    <NA>
#> 74         HS      Lei et al. 2019        <NA>    <NA>
#> 75         HS      Lei et al. 2019        <NA>    <NA>
#> 76         HS      Lei et al. 2019        <NA>    <NA>
#> 77         HS      Lei et al. 2019        <NA>    <NA>
#> 78         HS      Lei et al. 2019        <NA>    <NA>
#> 79         HS      Lei et al. 2019        <NA>    <NA>
#> 80         HS      Lei et al. 2019        <NA>    <NA>

query.citation("Lei et al. 2019", more.info = TRUE, ver = "upd")

#>      Category     Cas_No              Chemical_Name              Chemical_Name
#> 1      Ketone    67-64-1                    Acetone                    Acetone
#> 2      Ketone    67-64-1                    Acetone                    Acetone
#> 3      Ketone    67-64-1                    Acetone                    Acetone
#> 4      Ketone    67-64-1                    Acetone                    Acetone
#> 5      Ketone    67-64-1                    Acetone                    Acetone
#> 6      Ketone    67-64-1                    Acetone                    Acetone
#> 7      Ketone    67-64-1                    Acetone                    Acetone
#> 8      Ketone    67-64-1                    Acetone                    Acetone
#> 9     Alcohol    71-23-8                   Propanol                   Propanol
#> 10    Alcohol    71-23-8                   Propanol                   Propanol
#> 11    Alcohol    71-23-8                   Propanol                   Propanol
#> 12    Alcohol    71-23-8                   Propanol                   Propanol
#> 13    Alcohol    71-23-8                   Propanol                   Propanol
#> 14    Alcohol    71-23-8                   Propanol                   Propanol
#> 15    Alcohol    71-23-8                   Propanol                   Propanol
#> 16    Alcohol    71-23-8                   Propanol                   Propanol
#> 17    Alcohol    71-23-8                   Propanol                   Propanol
#> 18    Alcohol    71-36-3                    Butanol                    Butanol
#> 19    Alcohol    71-36-3                    Butanol                    Butanol
#> 20    Alcohol    71-36-3                    Butanol                    Butanol
#> 21    Alcohol    71-36-3                    Butanol                    Butanol
#> 22    Alcohol    71-36-3                    Butanol                    Butanol
#> 23    Alcohol    71-36-3                    Butanol                    Butanol
#> 24    Alcohol    71-36-3                    Butanol                    Butanol
#> 25    Alcohol    71-36-3                    Butanol                    Butanol
#> 26    Alcohol    71-36-3                    Butanol                    Butanol
#> 27    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 28    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 29    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 30    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 31    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 32    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 33    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 34    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 35    Nitrile 1975-05-08               Acetonitrile               Acetonitrile
#> 36      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 37      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 38      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 39      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 40      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 41      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 42      Arene   106-42-3                   p-Xylene                   p-Xylene
#> 43 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 44 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 45 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 46 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 47 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 48 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 49 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 50 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 51 Haloalkane   107-06-2         1,2-Dichloroethane         1,2-Dichloroethane
#> 52    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 53    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 54    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 55    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 56    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 57    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 58    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 59    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 60    Acetate   141-78-6              Ethyl acetate              Ethyl acetate
#> 61         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 62         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 63         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 64         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 65         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 66         CB   541-73-1        1,3-Dichlorobenzene        1,3-Dichlorobenzene
#> 67        PAH   493-02-7              trans-decalin              trans-decalin
#> 68        PAH   493-02-7              trans-decalin              trans-decalin
#> 69        PAH   493-02-7              trans-decalin              trans-decalin
#> 70        PAH   493-02-7              trans-decalin              trans-decalin
#> 71        PAH   493-02-7              trans-decalin              trans-decalin
#> 72        PAH   493-02-7              trans-decalin              trans-decalin
#> 73      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#> 74      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#> 75      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#> 76      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#> 77      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#> 78      Arene   488-23-3 1,2,3,4-Tetramethylbenzene 1,2,3,4-Tetramethylbenzene
#>                    IUPAC_name                              Alt_Chem_Name
#> 1                Propan-2-one                                  Propanone
#> 2                Propan-2-one                                  Propanone
#> 3                Propan-2-one                                  Propanone
#> 4                Propan-2-one                                  Propanone
#> 5                Propan-2-one                                  Propanone
#> 6                Propan-2-one                                  Propanone
#> 7                Propan-2-one                                  Propanone
#> 8                Propan-2-one                                  Propanone
#> 9                 Propan-1-ol                                 1-Propanol
#> 10                Propan-1-ol                                 1-Propanol
#> 11                Propan-1-ol                                 1-Propanol
#> 12                Propan-1-ol                                 1-Propanol
#> 13                Propan-1-ol                                 1-Propanol
#> 14                Propan-1-ol                                 1-Propanol
#> 15                Propan-1-ol                                 1-Propanol
#> 16                Propan-1-ol                                 1-Propanol
#> 17                Propan-1-ol                                 1-Propanol
#> 18                 Butan-1-ol                                  1-Butanol
#> 19                 Butan-1-ol                                  1-Butanol
#> 20                 Butan-1-ol                                  1-Butanol
#> 21                 Butan-1-ol                                  1-Butanol
#> 22                 Butan-1-ol                                  1-Butanol
#> 23                 Butan-1-ol                                  1-Butanol
#> 24                 Butan-1-ol                                  1-Butanol
#> 25                 Butan-1-ol                                  1-Butanol
#> 26                 Butan-1-ol                                  1-Butanol
#> 27               Acetonitrile                                       <NA>
#> 28               Acetonitrile                                       <NA>
#> 29               Acetonitrile                                       <NA>
#> 30               Acetonitrile                                       <NA>
#> 31               Acetonitrile                                       <NA>
#> 32               Acetonitrile                                       <NA>
#> 33               Acetonitrile                                       <NA>
#> 34               Acetonitrile                                       <NA>
#> 35               Acetonitrile                                       <NA>
#> 36                 1,4-Xylene                                       <NA>
#> 37                 1,4-Xylene                                       <NA>
#> 38                 1,4-Xylene                                       <NA>
#> 39                 1,4-Xylene                                       <NA>
#> 40                 1,4-Xylene                                       <NA>
#> 41                 1,4-Xylene                                       <NA>
#> 42                 1,4-Xylene                                       <NA>
#> 43         1,2-Dichloroethane                                       <NA>
#> 44         1,2-Dichloroethane                                       <NA>
#> 45         1,2-Dichloroethane                                       <NA>
#> 46         1,2-Dichloroethane                                       <NA>
#> 47         1,2-Dichloroethane                                       <NA>
#> 48         1,2-Dichloroethane                                       <NA>
#> 49         1,2-Dichloroethane                                       <NA>
#> 50         1,2-Dichloroethane                                       <NA>
#> 51         1,2-Dichloroethane                                       <NA>
#> 52              Ethyl acetate                    Acidic acid ethyl ester
#> 53              Ethyl acetate                    Acidic acid ethyl ester
#> 54              Ethyl acetate                    Acidic acid ethyl ester
#> 55              Ethyl acetate                    Acidic acid ethyl ester
#> 56              Ethyl acetate                    Acidic acid ethyl ester
#> 57              Ethyl acetate                    Acidic acid ethyl ester
#> 58              Ethyl acetate                    Acidic acid ethyl ester
#> 59              Ethyl acetate                    Acidic acid ethyl ester
#> 60              Ethyl acetate                    Acidic acid ethyl ester
#> 61        1,3-Dichlorobenzene                                       <NA>
#> 62        1,3-Dichlorobenzene                                       <NA>
#> 63        1,3-Dichlorobenzene                                       <NA>
#> 64        1,3-Dichlorobenzene                                       <NA>
#> 65        1,3-Dichlorobenzene                                       <NA>
#> 66        1,3-Dichlorobenzene                                       <NA>
#> 67 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 68 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 69 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 70 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 71 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 72 trans-Decahydronaphthalene 1,2,3,4,4a,5,6,7,8,8a-decahydronaphthalene
#> 73 1,2,3,4-Tetramethylbenzene                                       <NA>
#> 74 1,2,3,4-Tetramethylbenzene                                       <NA>
#> 75 1,2,3,4-Tetramethylbenzene                                       <NA>
#> 76 1,2,3,4-Tetramethylbenzene                                       <NA>
#> 77 1,2,3,4-Tetramethylbenzene                                       <NA>
#> 78 1,2,3,4-Tetramethylbenzene                                       <NA>
#>     Acronym Alt_Acronym Molar_Mass Temp  log_KOA Ranking Type Method_Type
#> 1      <NA>        <NA>      58.08   25 2.320000       4  EXP      Static
#> 2      <NA>        <NA>      58.08   37 2.190000       4  EXP      Static
#> 3      <NA>        <NA>      58.08   40 2.167317       4  EXP      Static
#> 4      <NA>        <NA>      58.08   50 2.025306       4  EXP      Static
#> 5      <NA>        <NA>      58.08   60 2.000000       4  EXP      Static
#> 6      <NA>        <NA>      58.08   80 1.792392       4  EXP      Static
#> 7      <NA>        <NA>      58.08   90 1.724276       4  EXP      Static
#> 8      <NA>        <NA>      58.08  100 1.623249       4  EXP      Static
#> 9      <NA>        <NA>     60.096   25 3.820000       4  EXP      Static
#> 10     <NA>        <NA>     60.096   37 3.520000       4  EXP      Static
#> 11     <NA>        <NA>     60.096   40 3.445915       4  EXP      Static
#> 12     <NA>        <NA>     60.096   50 3.221675       4  EXP      Static
#> 13     <NA>        <NA>     60.096   60 3.005609       4  EXP      Static
#> 14     <NA>        <NA>     60.096   70 2.784617       4  EXP      Static
#> 15     <NA>        <NA>     60.096   80 2.587711       4  EXP      Static
#> 16     <NA>        <NA>     60.096   90 2.418301       4  EXP      Static
#> 17     <NA>        <NA>     60.096  100 2.267172       4  EXP      Static
#> 18     <NA>        <NA>     74.123   25 4.100000       4  EXP      Static
#> 19     <NA>        <NA>     74.123   37 3.810000       4  EXP      Static
#> 20     <NA>        <NA>     74.123   40 3.643946       4  EXP      Static
#> 21     <NA>        <NA>     74.123   50 3.410946       4  EXP      Static
#> 22     <NA>        <NA>     74.123   60 3.317227       4  EXP      Static
#> 23     <NA>        <NA>     74.123   70 3.118595       4  EXP      Static
#> 24     <NA>        <NA>     74.123   80 2.938019       4  EXP      Static
#> 25     <NA>        <NA>     74.123   90 2.781755       4  EXP      Static
#> 26     <NA>        <NA>     74.123  100 2.595496       4  EXP      Static
#> 27     <NA>        <NA>     41.053   25 2.270000       4  EXP      Static
#> 28     <NA>        <NA>     41.053   37 2.140000       4  EXP      Static
#> 29     <NA>        <NA>     41.053   40 2.110590       4  EXP      Static
#> 30     <NA>        <NA>     41.053   50 1.968483       4  EXP      Static
#> 31     <NA>        <NA>     41.053   60 1.924279       4  EXP      Static
#> 32     <NA>        <NA>     41.053   70 1.851258       4  EXP      Static
#> 33     <NA>        <NA>     41.053   80 1.732394       4  EXP      Static
#> 34     <NA>        <NA>     41.053   90 1.633468       4  EXP      Static
#> 35     <NA>        <NA>     41.053  100 1.579784       4  EXP      Static
#> 36     <NA>        <NA>    106.168   25 3.590000       4  EXP      Static
#> 37     <NA>        <NA>    106.168   37 3.390000       4  EXP      Static
#> 38     <NA>        <NA>    106.168   70 2.885361       4  EXP      Static
#> 39     <NA>        <NA>    106.168   80 2.808211       4  EXP      Static
#> 40     <NA>        <NA>    106.168   90 2.673942       4  EXP      Static
#> 41     <NA>        <NA>    106.168  100 2.532754       4  EXP      Static
#> 42     <NA>        <NA>    106.168  110 2.457882       4  EXP      Static
#> 43     <NA>        <NA>      98.95   25 2.730000       4  EXP      Static
#> 44     <NA>        <NA>      98.95   37 2.570000       4  EXP      Static
#> 45     <NA>        <NA>      98.95   40 2.525045       4  EXP      Static
#> 46     <NA>        <NA>      98.95   50 2.380211       4  EXP      Static
#> 47     <NA>        <NA>      98.95   60 2.350248       4  EXP      Static
#> 48     <NA>        <NA>      98.95   70 2.206826       4  EXP      Static
#> 49     <NA>        <NA>      98.95   80 2.096910       4  EXP      Static
#> 50     <NA>        <NA>      98.95   90 1.973128       4  EXP      Static
#> 51     <NA>        <NA>      98.95  100 1.897627       4  EXP      Static
#> 52     <NA>        <NA>     88.106   25 2.630000       4  EXP      Static
#> 53     <NA>        <NA>     88.106   37 2.480000       4  EXP      Static
#> 54     <NA>        <NA>     88.106   40 2.401401       4  EXP      Static
#> 55     <NA>        <NA>     88.106   50 2.357935       4  EXP      Static
#> 56     <NA>        <NA>     88.106   60 2.305351       4  EXP      Static
#> 57     <NA>        <NA>     88.106   70 2.149219       4  EXP      Static
#> 58     <NA>        <NA>     88.106   80 2.049218       4  EXP      Static
#> 59     <NA>        <NA>     88.106   90 1.968483       4  EXP      Static
#> 60     <NA>        <NA>     88.106  100 1.875061       4  EXP      Static
#> 61 1,3-diCB        <NA>     147.01   25 4.170000       4  EXP      Static
#> 62 1,3-diCB        <NA>     147.01   37 3.940000       4  EXP      Static
#> 63 1,3-diCB        <NA>     147.01   80 3.247973       4  EXP      Static
#> 64 1,3-diCB        <NA>     147.01   90 3.111263       4  EXP      Static
#> 65 1,3-diCB        <NA>     147.01  100 2.992554       4  EXP      Static
#> 66 1,3-diCB        <NA>     147.01  110 2.845718       4  EXP      Static
#> 67     <NA>        <NA>    138.254   25 4.200000       4  EXP      Static
#> 68     <NA>        <NA>    138.254   37 3.980000       4  EXP      Static
#> 69     <NA>        <NA>    138.254   80 3.331225       4  EXP      Static
#> 70     <NA>        <NA>    138.254   90 3.177825       4  EXP      Static
#> 71     <NA>        <NA>    138.254  100 3.118926       4  EXP      Static
#> 72     <NA>        <NA>    138.254  110 2.938019       4  EXP      Static
#> 73     <NA>        <NA>    134.222   25 4.400000       4  EXP      Static
#> 74     <NA>        <NA>    134.222   37 4.300000       4  EXP      Static
#> 75     <NA>        <NA>    134.222   80 3.719331       4  EXP      Static
#> 76     <NA>        <NA>    134.222   90 3.568554       4  EXP      Static
#> 77     <NA>        <NA>    134.222  100 3.533009       4  EXP      Static
#> 78     <NA>        <NA>    134.222  110 3.378761       4  EXP      Static
#>    Gen_Method Method_Bin wet_dry_Octanol Value_Reported_As        Citation
#> 1        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 2        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 3        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 4        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 5        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 6        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 7        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 8        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 9        VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 10       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 11       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 12       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 13       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 14       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 15       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 16       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 17       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 18       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 19       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 20       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 21       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 22       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 23       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 24       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 25       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 26       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 27       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 28       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 29       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 30       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 31       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 32       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 33       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 34       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 35       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 36       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 37       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 38       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 39       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 40       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 41       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 42       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 43       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 44       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 45       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 46       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 47       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 48       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 49       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 50       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 51       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 52       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 53       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 54       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 55       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 56       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 57       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 58       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 59       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 60       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 61       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 62       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 63       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 64       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 65       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 66       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 67       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 68       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 69       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 70       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 71       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 72       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 73       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 74       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 75       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 76       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 77       VPHS         HS     dry octanol               KOA Lei et al. 2019
#> 78       VPHS         HS     dry octanol               KOA Lei et al. 2019
#>     Team Reliability Comment
#> 1  Wania        <NA>    <NA>
#> 2  Wania        <NA>    <NA>
#> 3  Wania        <NA>    <NA>
#> 4  Wania        <NA>    <NA>
#> 5  Wania        <NA>    <NA>
#> 6  Wania        <NA>    <NA>
#> 7  Wania        <NA>    <NA>
#> 8  Wania        <NA>    <NA>
#> 9  Wania        <NA>    <NA>
#> 10 Wania        <NA>    <NA>
#> 11 Wania        <NA>    <NA>
#> 12 Wania        <NA>    <NA>
#> 13 Wania        <NA>    <NA>
#> 14 Wania        <NA>    <NA>
#> 15 Wania        <NA>    <NA>
#> 16 Wania        <NA>    <NA>
#> 17 Wania        <NA>    <NA>
#> 18 Wania        <NA>    <NA>
#> 19 Wania        <NA>    <NA>
#> 20 Wania        <NA>    <NA>
#> 21 Wania        <NA>    <NA>
#> 22 Wania        <NA>    <NA>
#> 23 Wania        <NA>    <NA>
#> 24 Wania        <NA>    <NA>
#> 25 Wania        <NA>    <NA>
#> 26 Wania        <NA>    <NA>
#> 27 Wania        <NA>    <NA>
#> 28 Wania        <NA>    <NA>
#> 29 Wania        <NA>    <NA>
#> 30 Wania        <NA>    <NA>
#> 31 Wania        <NA>    <NA>
#> 32 Wania        <NA>    <NA>
#> 33 Wania        <NA>    <NA>
#> 34 Wania        <NA>    <NA>
#> 35 Wania        <NA>    <NA>
#> 36 Wania        <NA>    <NA>
#> 37 Wania        <NA>    <NA>
#> 38 Wania        <NA>    <NA>
#> 39 Wania        <NA>    <NA>
#> 40 Wania        <NA>    <NA>
#> 41 Wania        <NA>    <NA>
#> 42 Wania        <NA>    <NA>
#> 43 Wania        <NA>    <NA>
#> 44 Wania        <NA>    <NA>
#> 45 Wania        <NA>    <NA>
#> 46 Wania        <NA>    <NA>
#> 47 Wania        <NA>    <NA>
#> 48 Wania        <NA>    <NA>
#> 49 Wania        <NA>    <NA>
#> 50 Wania        <NA>    <NA>
#> 51 Wania        <NA>    <NA>
#> 52 Wania        <NA>    <NA>
#> 53 Wania        <NA>    <NA>
#> 54 Wania        <NA>    <NA>
#> 55 Wania        <NA>    <NA>
#> 56 Wania        <NA>    <NA>
#> 57 Wania        <NA>    <NA>
#> 58 Wania        <NA>    <NA>
#> 59 Wania        <NA>    <NA>
#> 60 Wania        <NA>    <NA>
#> 61 Wania        <NA>    <NA>
#> 62 Wania        <NA>    <NA>
#> 63 Wania        <NA>    <NA>
#> 64 Wania        <NA>    <NA>
#> 65 Wania        <NA>    <NA>
#> 66 Wania        <NA>    <NA>
#> 67 Wania        <NA>    <NA>
#> 68 Wania        <NA>    <NA>
#> 69 Wania        <NA>    <NA>
#> 70 Wania        <NA>    <NA>
#> 71 Wania        <NA>    <NA>
#> 72 Wania        <NA>    <NA>
#> 73 Wania        <NA>    <NA>
#> 74 Wania        <NA>    <NA>
#> 75 Wania        <NA>    <NA>
#> 76 Wania        <NA>    <NA>
#> 77 Wania        <NA>    <NA>
#> 78 Wania        <NA>    <NA>