Search by chemical name

Create a data frame with all estimated and experimental KOA values which exactly match the specified chemical name(s). All entries are case sensitive.

query.name(query, more.info = F, ver = "upd")

Arguments

query
Chemical name, either a single input or a vector. Searches against multiple synonymns included in the chemical table.

more.info
When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).

ver
Version of the database. Default is upd, for the latest version.

Value

A data frame containing KOA data for chemicals with names exactly matching the query.

See also

Examples

query.name("8:2 FTOH")

#>      Cas_No             Chemical_Name Temp  log_KOA Ranking Type Method_Bin
#> 1  678-39-7 8:2 Fluorotelomer alcohol    0 6.723000       4  EXP         FM
#> 2  678-39-7 8:2 Fluorotelomer alcohol    5 6.340000       4  EXP         FM
#> 3  678-39-7 8:2 Fluorotelomer alcohol   10 6.050000       4  EXP         FM
#> 4  678-39-7 8:2 Fluorotelomer alcohol   10 6.203000       4  EXP         FM
#> 5  678-39-7 8:2 Fluorotelomer alcohol   15 5.890000       4  EXP         FM
#> 6  678-39-7 8:2 Fluorotelomer alcohol   20 5.650000       4  EXP         FM
#> 7  678-39-7 8:2 Fluorotelomer alcohol   20 5.781000       4  EXP         FM
#> 8  678-39-7 8:2 Fluorotelomer alcohol   25 2.150000       3  EST  Solvation
#> 9  678-39-7 8:2 Fluorotelomer alcohol   25 4.170000       4  EXP         RT
#> 10 678-39-7 8:2 Fluorotelomer alcohol   25 4.216718       3  EST       QSPR
#> 11 678-39-7 8:2 Fluorotelomer alcohol   25 5.550000       4  EXP         FM
#> 12 678-39-7 8:2 Fluorotelomer alcohol   25 5.560000       4  EXP         FM
#> 13 678-39-7 8:2 Fluorotelomer alcohol   30 5.450000       4  EXP         FM
#> 14 678-39-7 8:2 Fluorotelomer alcohol   40 5.170000       4  EXP         FM
#>              Citation
#> 1    Goss et al. 2006
#> 2  Thuens et al. 2008
#> 3  Thuens et al. 2008
#> 4    Goss et al. 2006
#> 5  Thuens et al. 2008
#> 6  Thuens et al. 2008
#> 7    Goss et al. 2006
#> 8      Fu et al. 2016
#> 9     Lei et al. 2004
#> 10       Mathieu 2020
#> 11 Thuens et al. 2008
#> 12   Goss et al. 2006
#> 13 Thuens et al. 2008
#> 14 Thuens et al. 2008
#>                                                               Reliability
#> 1                                                                    <NA>
#> 2                                                                    <NA>
#> 3                                                                    <NA>
#> 4                                                                    <NA>
#> 5                                                                    <NA>
#> 6                                                                    <NA>
#> 7                                                                    <NA>
#> 8                                                                    <NA>
#> 9  chemical is too polar for methodology (Aa+Bb>2); A is  exp , B is  exp
#> 10                                                                   <NA>
#> 11                                                                   <NA>
#> 12                                                                   <NA>
#> 13                                                                   <NA>
#> 14                                                                   <NA>
#>    Comment
#> 1     <NA>
#> 2     <NA>
#> 3     <NA>
#> 4     <NA>
#> 5     <NA>
#> 6     <NA>
#> 7     <NA>
#> 8     <NA>
#> 9     <NA>
#> 10    <NA>
#> 11    <NA>
#> 12    <NA>
#> 13    <NA>
#> 14    <NA>

query.name("HCB", more.info = TRUE)

#>    Category   Cas_No     Chemical_Name     Chemical_Name        IUPAC_name
#> 1        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 2        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 3        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 4        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 5        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 6        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 7        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 8        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 9        CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 10       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 11       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 12       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 13       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 14       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 15       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 16       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 17       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 18       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 19       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 20       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 21       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 22       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 23       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 24       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 25       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 26       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 27       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 28       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 29       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 30       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 31       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#> 32       CB 118-74-1 Hexachlorobenzene Hexachlorobenzene Hexachlorobenzene
#>    Alt_Chem_Name Acronym Alt_Acronym Molar_Mass Temp  log_KOA Ranking Type
#> 1           <NA>     HCB        <NA>     284.77  -10 8.635361       4  EXP
#> 2           <NA>     HCB        <NA>     284.77    0 8.088278       4  EXP
#> 3           <NA>     HCB        <NA>     284.77    5 7.912302       3  EST
#> 4           <NA>     HCB        <NA>     284.77    5 8.104375       4  EXP
#> 5           <NA>     HCB        <NA>     284.77   10 7.579859       4  EXP
#> 6           <NA>     HCB        <NA>     284.77   10 7.772583       3  EST
#> 7           <NA>     HCB        <NA>     284.77   10 7.810000       3  EST
#> 8           <NA>     HCB        <NA>     284.77   10 7.850000       3  EST
#> 9           <NA>     HCB        <NA>     284.77   10 7.887005       4  EXP
#> 10          <NA>     HCB        <NA>     284.77   10 7.998000       3  EST
#> 11          <NA>     HCB        <NA>     284.77   15 7.637712       3  EST
#> 12          <NA>     HCB        <NA>     284.77   15 7.699948       4  EXP
#> 13          <NA>     HCB        <NA>     284.77   20 7.106143       4  EXP
#> 14          <NA>     HCB        <NA>     284.77   20 7.330000       3  EST
#> 15          <NA>     HCB        <NA>     284.77   20 7.507443       3  EST
#> 16          <NA>     HCB        <NA>     284.77   20 7.563416       4  EXP
#> 17          <NA>     HCB        <NA>     284.77   20 7.638000       3  EST
#> 18          <NA>     HCB        <NA>     284.77   25 4.930000       3  EST
#> 19          <NA>     HCB        <NA>     284.77   25 5.760000       3  EST
#> 20          <NA>     HCB        <NA>     284.77   25 6.890000       3  EST
#> 21          <NA>     HCB        <NA>     284.77   25 6.900000       4  EXP
#> 22          <NA>     HCB        <NA>     284.77   25 6.910000       3  EST
#> 23          <NA>     HCB        <NA>     284.77   25 6.910000       1  EST
#> 24          <NA>     HCB        <NA>     284.77   25 7.132230       3  EST
#> 25          <NA>     HCB        <NA>     284.77   25 7.170000       2  EST
#> 26          <NA>     HCB        <NA>     284.77   25 7.324670       2  EST
#> 27          <NA>     HCB        <NA>     284.77   25 7.381542       3  EST
#> 28          <NA>     HCB        <NA>     284.77   25 7.383523       4  EXP
#> 29          <NA>     HCB        <NA>     284.77   25 7.387799       4  EXP
#> 30          <NA>     HCB        <NA>     284.77   25 7.440000       3  EST
#> 31          <NA>     HCB        <NA>     284.77   25 7.880000       3  EST
#> 32          <NA>     HCB        <NA>     284.77   25 8.540000       1  EST
#>    Method_Type            Gen_Method Method_Bin wet_dry_Octanol
#> 1      Dynamic                    FM         FM     dry octanol
#> 2      Dynamic                    FM         FM     dry octanol
#> 3          EST              additive       QSPR     dry octanol
#> 4      Dynamic                    FM         FM     dry octanol
#> 5      Dynamic                    FM         FM     dry octanol
#> 6          EST              additive       QSPR     dry octanol
#> 7          EST                   MLR       QSPR     dry octanol
#> 8          EST              Triangle       QSPR     wet octanol
#> 9      Dynamic                    FM         FM     dry octanol
#> 10         EST                   MLR       QSPR     dry octanol
#> 11         EST              additive       QSPR     dry octanol
#> 12     Dynamic                    FM         FM     dry octanol
#> 13     Dynamic                    FM         FM     dry octanol
#> 14         EST                   MLR       QSPR     dry octanol
#> 15         EST              additive       QSPR     dry octanol
#> 16     Dynamic                    FM         FM     dry octanol
#> 17         EST                   MLR       QSPR     dry octanol
#> 18         EST SM8AD/CM4/M05-2X/6-31  Solvation     dry octanol
#> 19         EST                 UPPER       QSPR     dry octanol
#> 20         EST              Triangle       QSPR     wet octanol
#> 21     Dynamic                    FM         FM     dry octanol
#> 22         EST                   PLS       QSPR     dry octanol
#> 23         EST                   MLR       QSPR     dry octanol
#> 24         EST                  LFER       QSPR     dry octanol
#> 25         EST              Triangle       QSPR     dry octanol
#> 26         EST                ppLFER       QSPR     dry octanol
#> 27         EST              additive       QSPR     dry octanol
#> 28     Dynamic                    FM         FM     dry octanol
#> 29     Dynamic                    FM         FM     dry octanol
#> 30         EST                   MLR       QSPR     dry octanol
#> 31         EST              Triangle       QSPR     dry octanol
#> 32         EST              Triangle       QSPR     wet octanol
#>    Value_Reported_As                   Citation      Team
#> 1            log KOA     Harner and Mackay 1995    Harner
#> 2            log KOA     Harner and Mackay 1995    Harner
#> 3            log KOA               Mathieu 2020   Mathieu
#> 4            log KOA    Shoeib and Harner 2002b    Harner
#> 5            log KOA     Harner and Mackay 1995    Harner
#> 6            log KOA               Mathieu 2020   Mathieu
#> 7            log KOA            Jin et al. 2017      Chen
#> 8            log KOA     Meylan and Howard 2005    Meylan
#> 9            log KOA    Shoeib and Harner 2002b    Harner
#> 10           log KOA            Li et al. 2006a      Chen
#> 11           log KOA               Mathieu 2020   Mathieu
#> 12           log KOA    Shoeib and Harner 2002b    Harner
#> 13           log KOA     Harner and Mackay 1995    Harner
#> 14           log KOA            Jin et al. 2017      Chen
#> 15           log KOA               Mathieu 2020   Mathieu
#> 16           log KOA    Shoeib and Harner 2002b    Harner
#> 17           log KOA            Li et al. 2006a      Chen
#> 18           log KOA             Fu et al. 2016      Chen
#> 19           log KAO      Yalkowsky et al. 1994 Yalkowsky
#> 20           log KOA     Meylan and Howard 2005    Meylan
#> 21           log KOA     Harner and Mackay 1995    Harner
#> 22           log KOA          Chen et al. 2003c      Chen
#> 23           log KOA           Zhao et al. 2005     Liang
#> 24           log KOA           Nabi et al. 2014      Nabi
#> 25           log KOA        Abraham et al. 2001   Abraham
#> 26           log KOA           Nabi et al. 2014      Nabi
#> 27           log KOA               Mathieu 2020   Mathieu
#> 28           log KOA    Shoeib and Harner 2002b    Harner
#> 29           log KOA    Shoeib and Harner 2002b    Harner
#> 30           log KOA            Li et al. 2006a      Chen
#> 31           log KOA Sepassi and Yalkowsky 2007   Sepassi
#> 32           log KOA        Finizio et al. 1997    Harner
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                                                                                     <NA>
#> 10                                                                                                                                    <NA>
#> 11                                                                                                                                    <NA>
#> 12                                                                                                                                    <NA>
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#> 22                                                                                                                                    <NA>
#> 23                                                                                                                                    <NA>
#> 24                                                                                                                                    <NA>
#> 25                                                                                                                                    <NA>
#> 26                                                                                                                                    <NA>
#> 27                                                                                                                                    <NA>
#> 28                                                                                                  Reported in paper as recommended value
#> 29 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 30                                                                                                                                    <NA>
#> 31                                                                                                                                    <NA>
#> 32                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                Raw data and errors from communication with T. Harner. 
#> 5                                                                   <NA>
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                Raw data and errors from communication with T. Harner. 
#> 10                                                                  <NA>
#> 11                                                                  <NA>
#> 12               Raw data and errors from communication with T. Harner. 
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16               Raw data and errors from communication with T. Harner. 
#> 17                                                                  <NA>
#> 18                                                                  <NA>
#> 19                                                                  <NA>
#> 20                                                                  <NA>
#> 21                                                                  <NA>
#> 22                                                                  <NA>
#> 23                                                                  <NA>
#> 24                                                                  <NA>
#> 25                                                                  <NA>
#> 26                                                                  <NA>
#> 27                                                                  <NA>
#> 28 Calculated from regression equation and originally reported in paper.
#> 29               Raw data and errors from communication with T. Harner. 
#> 30                                                                  <NA>
#> 31                                                                  <NA>
#> 32                                                                  <NA>

query.name(c("8:2 FTOH", "Hexachlorobenzene"))

#>      Cas_No             Chemical_Name Temp  log_KOA Ranking Type Method_Bin
#> 1  678-39-7 8:2 Fluorotelomer alcohol    0 6.723000       4  EXP         FM
#> 2  678-39-7 8:2 Fluorotelomer alcohol    5 6.340000       4  EXP         FM
#> 3  678-39-7 8:2 Fluorotelomer alcohol   10 6.050000       4  EXP         FM
#> 4  678-39-7 8:2 Fluorotelomer alcohol   10 6.203000       4  EXP         FM
#> 5  678-39-7 8:2 Fluorotelomer alcohol   15 5.890000       4  EXP         FM
#> 6  678-39-7 8:2 Fluorotelomer alcohol   20 5.650000       4  EXP         FM
#> 7  678-39-7 8:2 Fluorotelomer alcohol   20 5.781000       4  EXP         FM
#> 8  678-39-7 8:2 Fluorotelomer alcohol   25 2.150000       3  EST  Solvation
#> 9  678-39-7 8:2 Fluorotelomer alcohol   25 4.170000       4  EXP         RT
#> 10 678-39-7 8:2 Fluorotelomer alcohol   25 4.216718       3  EST       QSPR
#> 11 678-39-7 8:2 Fluorotelomer alcohol   25 5.550000       4  EXP         FM
#> 12 678-39-7 8:2 Fluorotelomer alcohol   25 5.560000       4  EXP         FM
#> 13 678-39-7 8:2 Fluorotelomer alcohol   30 5.450000       4  EXP         FM
#> 14 678-39-7 8:2 Fluorotelomer alcohol   40 5.170000       4  EXP         FM
#> 15 118-74-1         Hexachlorobenzene  -10 8.635361       4  EXP         FM
#> 16 118-74-1         Hexachlorobenzene    0 8.088278       4  EXP         FM
#> 17 118-74-1         Hexachlorobenzene    5 7.912302       3  EST       QSPR
#> 18 118-74-1         Hexachlorobenzene    5 8.104375       4  EXP         FM
#> 19 118-74-1         Hexachlorobenzene   10 7.579859       4  EXP         FM
#> 20 118-74-1         Hexachlorobenzene   10 7.772583       3  EST       QSPR
#> 21 118-74-1         Hexachlorobenzene   10 7.810000       3  EST       QSPR
#> 22 118-74-1         Hexachlorobenzene   10 7.850000       3  EST       QSPR
#> 23 118-74-1         Hexachlorobenzene   10 7.887005       4  EXP         FM
#> 24 118-74-1         Hexachlorobenzene   10 7.998000       3  EST       QSPR
#> 25 118-74-1         Hexachlorobenzene   15 7.637712       3  EST       QSPR
#> 26 118-74-1         Hexachlorobenzene   15 7.699948       4  EXP         FM
#> 27 118-74-1         Hexachlorobenzene   20 7.106143       4  EXP         FM
#> 28 118-74-1         Hexachlorobenzene   20 7.330000       3  EST       QSPR
#> 29 118-74-1         Hexachlorobenzene   20 7.507443       3  EST       QSPR
#> 30 118-74-1         Hexachlorobenzene   20 7.563416       4  EXP         FM
#> 31 118-74-1         Hexachlorobenzene   20 7.638000       3  EST       QSPR
#> 32 118-74-1         Hexachlorobenzene   25 4.930000       3  EST  Solvation
#> 33 118-74-1         Hexachlorobenzene   25 5.760000       3  EST       QSPR
#> 34 118-74-1         Hexachlorobenzene   25 6.890000       3  EST       QSPR
#> 35 118-74-1         Hexachlorobenzene   25 6.900000       4  EXP         FM
#> 36 118-74-1         Hexachlorobenzene   25 6.910000       3  EST       QSPR
#> 37 118-74-1         Hexachlorobenzene   25 6.910000       1  EST       QSPR
#> 38 118-74-1         Hexachlorobenzene   25 7.132230       3  EST       QSPR
#> 39 118-74-1         Hexachlorobenzene   25 7.170000       2  EST       QSPR
#> 40 118-74-1         Hexachlorobenzene   25 7.324670       2  EST       QSPR
#> 41 118-74-1         Hexachlorobenzene   25 7.381542       3  EST       QSPR
#> 42 118-74-1         Hexachlorobenzene   25 7.383523       4  EXP         FM
#> 43 118-74-1         Hexachlorobenzene   25 7.387799       4  EXP         FM
#> 44 118-74-1         Hexachlorobenzene   25 7.440000       3  EST       QSPR
#> 45 118-74-1         Hexachlorobenzene   25 7.880000       3  EST       QSPR
#> 46 118-74-1         Hexachlorobenzene   25 8.540000       1  EST       QSPR
#>                      Citation
#> 1            Goss et al. 2006
#> 2          Thuens et al. 2008
#> 3          Thuens et al. 2008
#> 4            Goss et al. 2006
#> 5          Thuens et al. 2008
#> 6          Thuens et al. 2008
#> 7            Goss et al. 2006
#> 8              Fu et al. 2016
#> 9             Lei et al. 2004
#> 10               Mathieu 2020
#> 11         Thuens et al. 2008
#> 12           Goss et al. 2006
#> 13         Thuens et al. 2008
#> 14         Thuens et al. 2008
#> 15     Harner and Mackay 1995
#> 16     Harner and Mackay 1995
#> 17               Mathieu 2020
#> 18    Shoeib and Harner 2002b
#> 19     Harner and Mackay 1995
#> 20               Mathieu 2020
#> 21            Jin et al. 2017
#> 22     Meylan and Howard 2005
#> 23    Shoeib and Harner 2002b
#> 24            Li et al. 2006a
#> 25               Mathieu 2020
#> 26    Shoeib and Harner 2002b
#> 27     Harner and Mackay 1995
#> 28            Jin et al. 2017
#> 29               Mathieu 2020
#> 30    Shoeib and Harner 2002b
#> 31            Li et al. 2006a
#> 32             Fu et al. 2016
#> 33      Yalkowsky et al. 1994
#> 34     Meylan and Howard 2005
#> 35     Harner and Mackay 1995
#> 36          Chen et al. 2003c
#> 37           Zhao et al. 2005
#> 38           Nabi et al. 2014
#> 39        Abraham et al. 2001
#> 40           Nabi et al. 2014
#> 41               Mathieu 2020
#> 42    Shoeib and Harner 2002b
#> 43    Shoeib and Harner 2002b
#> 44            Li et al. 2006a
#> 45 Sepassi and Yalkowsky 2007
#> 46        Finizio et al. 1997
#>                                                                                                                                Reliability
#> 1                                                                                                                                     <NA>
#> 2                                                                                                                                     <NA>
#> 3                                                                                                                                     <NA>
#> 4                                                                                                                                     <NA>
#> 5                                                                                                                                     <NA>
#> 6                                                                                                                                     <NA>
#> 7                                                                                                                                     <NA>
#> 8                                                                                                                                     <NA>
#> 9                                                                   chemical is too polar for methodology (Aa+Bb>2); A is  exp , B is  exp
#> 10                                                                                                                                    <NA>
#> 11                                                                                                                                    <NA>
#> 12                                                                                                                                    <NA>
#> 13                                                                                                                                    <NA>
#> 14                                                                                                                                    <NA>
#> 15                                                                                                                                    <NA>
#> 16                                                                                                                                    <NA>
#> 17                                                                                                                                    <NA>
#> 18                                                                                                                                    <NA>
#> 19                                                                                                                                    <NA>
#> 20                                                                                                                                    <NA>
#> 21                                                                                                                                    <NA>
#> 22                                                                                                                                    <NA>
#> 23                                                                                                                                    <NA>
#> 24                                                                                                                                    <NA>
#> 25                                                                                                                                    <NA>
#> 26                                                                                                                                    <NA>
#> 27                                                                                                                                    <NA>
#> 28                                                                                                                                    <NA>
#> 29                                                                                                                                    <NA>
#> 30                                                                                                                                    <NA>
#> 31                                                                                                                                    <NA>
#> 32                                                                                                                                    <NA>
#> 33                                                                                                                                    <NA>
#> 34                                                                                                                                    <NA>
#> 35                                                                                                                                    <NA>
#> 36                                                                                                                                    <NA>
#> 37                                                                                                                                    <NA>
#> 38                                                                                                                                    <NA>
#> 39                                                                                                                                    <NA>
#> 40                                                                                                                                    <NA>
#> 41                                                                                                                                    <NA>
#> 42                                                                                                  Reported in paper as recommended value
#> 43 This value does not match the paper because the reported value is calculated from regression analysis reported/recommended by authors. 
#> 44                                                                                                                                    <NA>
#> 45                                                                                                                                    <NA>
#> 46                                                                                                                                    <NA>
#>                                                                  Comment
#> 1                                                                   <NA>
#> 2                                                                   <NA>
#> 3                                                                   <NA>
#> 4                                                                   <NA>
#> 5                                                                   <NA>
#> 6                                                                   <NA>
#> 7                                                                   <NA>
#> 8                                                                   <NA>
#> 9                                                                   <NA>
#> 10                                                                  <NA>
#> 11                                                                  <NA>
#> 12                                                                  <NA>
#> 13                                                                  <NA>
#> 14                                                                  <NA>
#> 15                                                                  <NA>
#> 16                                                                  <NA>
#> 17                                                                  <NA>
#> 18               Raw data and errors from communication with T. Harner. 
#> 19                                                                  <NA>
#> 20                                                                  <NA>
#> 21                                                                  <NA>
#> 22                                                                  <NA>
#> 23               Raw data and errors from communication with T. Harner. 
#> 24                                                                  <NA>
#> 25                                                                  <NA>
#> 26               Raw data and errors from communication with T. Harner. 
#> 27                                                                  <NA>
#> 28                                                                  <NA>
#> 29                                                                  <NA>
#> 30               Raw data and errors from communication with T. Harner. 
#> 31                                                                  <NA>
#> 32                                                                  <NA>
#> 33                                                                  <NA>
#> 34                                                                  <NA>
#> 35                                                                  <NA>
#> 36                                                                  <NA>
#> 37                                                                  <NA>
#> 38                                                                  <NA>
#> 39                                                                  <NA>
#> 40                                                                  <NA>
#> 41                                                                  <NA>
#> 42 Calculated from regression equation and originally reported in paper.
#> 43               Raw data and errors from communication with T. Harner. 
#> 44                                                                  <NA>
#> 45                                                                  <NA>
#> 46                                                                  <NA>

query.name("8:2 FTOH", ver = "upd")

#>      Cas_No             Chemical_Name Temp  log_KOA Ranking Type Method_Bin
#> 1  678-39-7 8:2 Fluorotelomer alcohol    0 6.723000       4  EXP         FM
#> 2  678-39-7 8:2 Fluorotelomer alcohol    5 6.340000       4  EXP         FM
#> 3  678-39-7 8:2 Fluorotelomer alcohol   10 6.050000       4  EXP         FM
#> 4  678-39-7 8:2 Fluorotelomer alcohol   10 6.203000       4  EXP         FM
#> 5  678-39-7 8:2 Fluorotelomer alcohol   15 5.890000       4  EXP         FM
#> 6  678-39-7 8:2 Fluorotelomer alcohol   20 5.650000       4  EXP         FM
#> 7  678-39-7 8:2 Fluorotelomer alcohol   20 5.781000       4  EXP         FM
#> 8  678-39-7 8:2 Fluorotelomer alcohol   25 2.150000       3  EST  Solvation
#> 9  678-39-7 8:2 Fluorotelomer alcohol   25 4.170000       4  EXP         RT
#> 10 678-39-7 8:2 Fluorotelomer alcohol   25 4.216718       3  EST       QSPR
#> 11 678-39-7 8:2 Fluorotelomer alcohol   25 5.550000       4  EXP         FM
#> 12 678-39-7 8:2 Fluorotelomer alcohol   25 5.560000       4  EXP         FM
#> 13 678-39-7 8:2 Fluorotelomer alcohol   30 5.450000       4  EXP         FM
#> 14 678-39-7 8:2 Fluorotelomer alcohol   40 5.170000       4  EXP         FM
#>              Citation
#> 1    Goss et al. 2006
#> 2  Thuens et al. 2008
#> 3  Thuens et al. 2008
#> 4    Goss et al. 2006
#> 5  Thuens et al. 2008
#> 6  Thuens et al. 2008
#> 7    Goss et al. 2006
#> 8      Fu et al. 2016
#> 9     Lei et al. 2004
#> 10       Mathieu 2020
#> 11 Thuens et al. 2008
#> 12   Goss et al. 2006
#> 13 Thuens et al. 2008
#> 14 Thuens et al. 2008
#>                                                               Reliability
#> 1                                                                    <NA>
#> 2                                                                    <NA>
#> 3                                                                    <NA>
#> 4                                                                    <NA>
#> 5                                                                    <NA>
#> 6                                                                    <NA>
#> 7                                                                    <NA>
#> 8                                                                    <NA>
#> 9  chemical is too polar for methodology (Aa+Bb>2); A is  exp , B is  exp
#> 10                                                                   <NA>
#> 11                                                                   <NA>
#> 12                                                                   <NA>
#> 13                                                                   <NA>
#> 14                                                                   <NA>
#>    Comment
#> 1     <NA>
#> 2     <NA>
#> 3     <NA>
#> 4     <NA>
#> 5     <NA>
#> 6     <NA>
#> 7     <NA>
#> 8     <NA>
#> 9     <NA>
#> 10    <NA>
#> 11    <NA>
#> 12    <NA>
#> 13    <NA>
#> 14    <NA>